Function syntax::algo::diff

pub fn diff(from: &SyntaxNode, to: &SyntaxNode) -> TreeDiff

Finds a (potentially minimal) diff, which, applied to from, will result in to.

Specifically, returns a structure that consists of a replacements, insertions and deletions such that applying this map on from will result in to.

This function tries to find a fine-grained diff.