Enum syntax::ast::Expr

pub enum Expr {

    ArrayExpr(ArrayExpr),
    AwaitExpr(AwaitExpr),
    BinExpr(BinExpr),
    BlockExpr(BlockExpr),
    BoxExpr(BoxExpr),
    BreakExpr(BreakExpr),
    CallExpr(CallExpr),
    CastExpr(CastExpr),
    ClosureExpr(ClosureExpr),
    ContinueExpr(ContinueExpr),
    FieldExpr(FieldExpr),
    ForExpr(ForExpr),
    IfExpr(IfExpr),
    IndexExpr(IndexExpr),
    Literal(Literal),
    LoopExpr(LoopExpr),
    MacroExpr(MacroExpr),
    MatchExpr(MatchExpr),
    MethodCallExpr(MethodCallExpr),
    ParenExpr(ParenExpr),
    PathExpr(PathExpr),
    PrefixExpr(PrefixExpr),
    RangeExpr(RangeExpr),
    RecordExpr(RecordExpr),
    RefExpr(RefExpr),
    ReturnExpr(ReturnExpr),
    TryExpr(TryExpr),
    TupleExpr(TupleExpr),
    WhileExpr(WhileExpr),
    YieldExpr(YieldExpr),
    YeetExpr(YeetExpr),
    LetExpr(LetExpr),
    UnderscoreExpr(UnderscoreExpr),

}

Variants

ArrayExpr(ArrayExpr)

AwaitExpr(AwaitExpr)

BinExpr(BinExpr)

BlockExpr(BlockExpr)

BoxExpr(BoxExpr)

BreakExpr(BreakExpr)

CallExpr(CallExpr)

CastExpr(CastExpr)

ClosureExpr(ClosureExpr)

ContinueExpr(ContinueExpr)

FieldExpr(FieldExpr)

ForExpr(ForExpr)

IfExpr(IfExpr)

IndexExpr(IndexExpr)

Literal(Literal)

LoopExpr(LoopExpr)

MacroExpr(MacroExpr)

MatchExpr(MatchExpr)

MethodCallExpr(MethodCallExpr)

ParenExpr(ParenExpr)

PathExpr(PathExpr)

PrefixExpr(PrefixExpr)

RangeExpr(RangeExpr)

RecordExpr(RecordExpr)

RefExpr(RefExpr)

ReturnExpr(ReturnExpr)

TryExpr(TryExpr)

TupleExpr(TupleExpr)

WhileExpr(WhileExpr)

YieldExpr(YieldExpr)

YeetExpr(YeetExpr)

LetExpr(LetExpr)

UnderscoreExpr(UnderscoreExpr)

Implementations

impl Expr

pub fn is_block_like(&self) -> bool

impl Expr

pub fn needs_parens_in(&self, parent: SyntaxNode) -> bool

Returns true if self would need to be wrapped in parentheses given that its parent is parent.