Enum syntax::ast::ExprKind

pub enum ExprKind {
    Box(P<Expr>),
    InPlace(P<Expr>, P<Expr>),
    Vec(Vec<P<Expr>>),
    Call(P<Expr>, Vec<P<Expr>>),
    MethodCall(SpannedIdent, Vec<P<Ty>>, Vec<P<Expr>>),
    Tup(Vec<P<Expr>>),
    Binary(BinOp, P<Expr>, P<Expr>),
    Unary(UnOp, P<Expr>),
    Lit(P<Lit>),
    Cast(P<Expr>, P<Ty>),
    Type(P<Expr>, P<Ty>),
    If(P<Expr>, P<Block>, Option<P<Expr>>),
    IfLet(P<Pat>, P<Expr>, P<Block>, Option<P<Expr>>),
    While(P<Expr>, P<Block>, Option<Ident>),
    WhileLet(P<Pat>, P<Expr>, P<Block>, Option<Ident>),
    ForLoop(P<Pat>, P<Expr>, P<Block>, Option<Ident>),
    Loop(P<Block>, Option<Ident>),
    Match(P<Expr>, Vec<Arm>),
    Closure(CaptureBy, P<FnDecl>, P<Block>),
    Block(P<Block>),
    Assign(P<Expr>, P<Expr>),
    AssignOp(BinOp, P<Expr>, P<Expr>),
    Field(P<Expr>, SpannedIdent),
    TupField(P<Expr>, Spanned<usize>),
    Index(P<Expr>, P<Expr>),
    Range(Option<P<Expr>>, Option<P<Expr>>),
    Path(Option<QSelf>, Path),
    AddrOf(Mutability, P<Expr>),
    Break(Option<SpannedIdent>),
    Again(Option<SpannedIdent>),
    Ret(Option<P<Expr>>),
    InlineAsm(InlineAsm),
    Mac(Mac),
    Struct(Path, Vec<Field>, Option<P<Expr>>),
    Repeat(P<Expr>, P<Expr>),
    Paren(P<Expr>),
}

Unstable (rustc_private)

Variants

Box
Unstable (rustc_private)

A box x expression.

InPlace
Unstable (rustc_private)

First expr is the place; second expr is the value.

Vec
Unstable (rustc_private)

An array ([a, b, c, d])

Call
Unstable (rustc_private)

A function call

The first field resolves to the function itself, and the second field is the list of arguments

MethodCall
Unstable (rustc_private)

A method call (x.foo::<Bar, Baz>(a, b, c, d))

The SpannedIdent is the identifier for the method name. The vector of Tys are the ascripted type parameters for the method (within the angle brackets).

The first element of the vector of Exprs is the expression that evaluates to the object on which the method is being called on (the receiver), and the remaining elements are the rest of the arguments.

Thus, x.foo::<Bar, Baz>(a, b, c, d) is represented as ExprKind::MethodCall(foo, [Bar, Baz], [x, a, b, c, d]).

Tup
Unstable (rustc_private)

A tuple ((a, b, c ,d))

Binary
Unstable (rustc_private)

A binary operation (For example: a + b, a * b)

Unary
Unstable (rustc_private)

A unary operation (For example: !x, *x)

Lit
Unstable (rustc_private)

A literal (For example: 1u8, "foo")

Cast
Unstable (rustc_private)

A cast (foo as f64)

Type
Unstable (rustc_private)

If
Unstable (rustc_private)

An if block, with an optional else block

if expr { block } else { expr }

IfLet
Unstable (rustc_private)

An if let expression with an optional else block

if let pat = expr { block } else { expr }

This is desugared to a match expression.

While
Unstable (rustc_private)

A while loop, with an optional label

'label: while expr { block }

WhileLet
Unstable (rustc_private)

A while-let loop, with an optional label

'label: while let pat = expr { block }

This is desugared to a combination of loop and match expressions.

ForLoop
Unstable (rustc_private)

A for loop, with an optional label

'label: for pat in expr { block }

This is desugared to a combination of loop and match expressions.

Loop
Unstable (rustc_private)

Conditionless loop (can be exited with break, continue, or return)

'label: loop { block }

Match
Unstable (rustc_private)

A match block.

Closure
Unstable (rustc_private)

A closure (for example, move |a, b, c| {a + b + c})

Block
Unstable (rustc_private)

A block ({ ... })

Assign
Unstable (rustc_private)

An assignment (a = foo())

AssignOp
Unstable (rustc_private)

An assignment with an operator

For example, a += 1.

Field
Unstable (rustc_private)

Access of a named struct field (obj.foo)

TupField
Unstable (rustc_private)

Access of an unnamed field of a struct or tuple-struct

For example, foo.0.

Index
Unstable (rustc_private)

An indexing operation (foo[2])

Range
Unstable (rustc_private)

A range (1..2, 1.., or ..2)

Path
Unstable (rustc_private)

Variable reference, possibly containing :: and/or type parameters, e.g. foo::bar::.

Optionally "qualified", e.g. <Vec<T> as SomeTrait>::SomeType.

AddrOf
Unstable (rustc_private)

A referencing operation (&a or &mut a)

Break
Unstable (rustc_private)

A break, with an optional label to break

Again
Unstable (rustc_private)

A continue, with an optional label

Ret
Unstable (rustc_private)

A return, with an optional value to be returned

InlineAsm
Unstable (rustc_private)

Output of the asm!() macro

Mac
Unstable (rustc_private)

A macro invocation; pre-expansion

Struct
Unstable (rustc_private)

A struct literal expression.

For example, Foo {x: 1, y: 2}, or Foo {x: 1, .. base}, where base is the Option<Expr>.

Repeat
Unstable (rustc_private)

An array literal constructed from one repeated element.

For example, [1u8; 5]. The first expression is the element to be repeated; the second is the number of times to repeat it.

Paren
Unstable (rustc_private)

No-op: used solely so we can pretty-print faithfully