Enum syntax::ast::StmtKind

pub enum StmtKind {
    Local(P<Local>),
    Item(P<Item>),
    Expr(P<Expr>),
    Semi(P<Expr>),
    Mac(P<(Mac, MacStmtStyle, ThinVec<Attribute>)>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
Local(P<Local>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A local (let) binding.

Item(P<Item>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An item definition.

Expr(P<Expr>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Expr without trailing semi-colon.

Semi(P<Expr>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Expr with a trailing semi-colon.

Mac(P<(Mac, MacStmtStyle, ThinVec<Attribute>)>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Macro.