syntax::ast

Module node_ext

Various extension methods to ast Nodes, which are hard to code-generate. Extensions for various expressions live in a sibling expr_extensions module.

These methods should only do simple, shallow tasks related to the syntax of the node itself.

Structs

SlicePatComponents

TokenTreeChildren

An iterator over the elements in an ast::TokenTree.

Enums

AttrKind

FieldKind

Macro

NameLike

NameOrNameRef

PathSegmentKind

SelfParamKind

StructKind

TypeBoundKind

TypeOrConstParam

VisibilityKind

Functions

text_of_first_token 🔒