Struct syntax::ast::TraitItem

pub struct TraitItem {
    pub id: NodeId,
    pub ident: Ident,
    pub attrs: Vec<Attribute>,
    pub node: TraitItem_,
    pub span: Span,
}

Unstable

Represents a method declaration in a trait declaration, possibly including a default implementation A trait method is either required (meaning it doesn't have an implementation, just a signature) or provided (meaning it has a default implementation).

Fields

id
Unstable

ident
Unstable

attrs
Unstable

node
Unstable

span
Unstable