syntax::ast::syntax_factory
Struct SyntaxFactory

pub struct SyntaxFactory { /* private fields */ }

Implementations

impl SyntaxFactory

pub fn name(&self, name: &str) -> Name

pub fn ty(&self, text: &str) -> Type

pub fn ty_infer(&self) -> InferType

pub fn type_param(&self, name: Name, bounds: Option<TypeBoundList>) -> TypeParam

pub fn ident_pat(&self, ref_: bool, mut_: bool, name: Name) -> IdentPat

pub fn block_expr( &self, stmts: impl IntoIterator<Item = Stmt>, tail_expr: Option<Expr>, ) -> BlockExpr

pub fn expr_empty_block(&self) -> BlockExpr

pub fn expr_bin(&self, lhs: Expr, op: BinaryOp, rhs: Expr) -> BinExpr

pub fn expr_path(&self, path: Path) -> Expr

pub fn expr_ref(&self, expr: Expr, exclusive: bool) -> Expr

pub fn expr_return(&self, expr: Option<Expr>) -> ReturnExpr

pub fn let_stmt( &self, pattern: Pat, ty: Option<Type>, initializer: Option<Expr>, ) -> LetStmt

pub fn item_const( &self, visibility: Option<Visibility>, name: Name, ty: Type, expr: Expr, ) -> Const

pub fn item_static( &self, visibility: Option<Visibility>, is_unsafe: bool, is_mut: bool, name: Name, ty: Type, expr: Option<Expr>, ) -> Static

pub fn turbofish_generic_arg_list( &self, args: impl IntoIterator<Item = GenericArg> + Clone, ) -> GenericArgList

pub fn record_field_list( &self, fields: impl IntoIterator<Item = RecordField>, ) -> RecordFieldList

pub fn record_field( &self, visibility: Option<Visibility>, name: Name, ty: Type, ) -> RecordField

pub fn tuple_field_list( &self, fields: impl IntoIterator<Item = TupleField>, ) -> TupleFieldList

pub fn tuple_field( &self, visibility: Option<Visibility>, ty: Type, ) -> TupleField

pub fn item_enum( &self, visibility: Option<Visibility>, name: Name, generic_param_list: Option<GenericParamList>, where_clause: Option<WhereClause>, variant_list: VariantList, ) -> Enum

pub fn variant_list( &self, variants: impl IntoIterator<Item = Variant>, ) -> VariantList

pub fn variant( &self, visibility: Option<Visibility>, name: Name, field_list: Option<FieldList>, discriminant: Option<Expr>, ) -> Variant

pub fn token_tree( &self, delimiter: SyntaxKind, tt: Vec<NodeOrToken<TokenTree, SyntaxToken>>, ) -> TokenTree

pub fn token(&self, kind: SyntaxKind) -> SyntaxToken

pub fn whitespace(&self, text: &str) -> SyntaxToken

impl SyntaxFactory

pub fn new() -> Self

Creates a new SyntaxFactory, generating mappings between input nodes and generated nodes.

pub fn without_mappings() -> Self

Creates a SyntaxFactory without generating mappings.

pub fn finish_with_mappings(self) -> SyntaxMapping

Gets all of the tracked syntax mappings, if any.