syntax::ast

Trait HasAttrs

pub trait HasAttrs: AstNode {
    // Provided methods
    fn attrs(&self) -> AstChildren<Attr>  { ... }
    fn inner_attributes_node(&self) -> Option<SyntaxNode> { ... }
}

Provided Methods

fn attrs(&self) -> AstChildren<Attr>

fn inner_attributes_node(&self) -> Option<SyntaxNode>

This may return the same node as called with (with SourceFile). The caller has the responsibility to avoid duplicate attributes.