Trait syntax::ast::Phase

pub trait Phasewhere
    Self: Default + Clone + Debug + Eq,{
    type Info: HasSpan + Clone + Debug;
    type TypeInfo: HasSpan + Clone + Debug;
    type TypeAppInfo: HasSpan + Clone + Into<Self::TypeInfo> + Debug;
    type VarName: Clone + Debug;
    type Typ: Clone + Debug;

    fn print_var(name: &Self::VarName, idx: Option<Idx>) -> String;
}

Required Associated Types

type Info: HasSpan + Clone + Debug

Type of the info field, containing span information

type TypeInfo: HasSpan + Clone + Debug

Type of the info field, containing span and (depending on the phase) type information

type TypeAppInfo: HasSpan + Clone + Into<Self::TypeInfo> + Debug

Type of the info field, containing span and (depending on the phase) type information where the type is required to be the full application of a type constructor

type VarName: Clone + Debug

Type of the name field of Exp::Var

type Typ: Clone + Debug

Type of the typ field for ParamInst

Required Methods

fn print_var(name: &Self::VarName, idx: Option<Idx>) -> String