Module syntax::ast_util

Unstable

Structs

IdRange 	 [Unstable]

IdVisitor 	 [Unstable]
A visitor that applies its operation to all of the node IDs in a visitable thing.

Constants

AS_PREC 	 [Unstable]
Precedence of the as operator, which is a binary operator not appearing in the prior table.

Traits

IdVisitingOperation 	 [Unstable]

Functions

binop_to_string 	 [Unstable]

compute_id_range_for_fn_body 	 [Unstable]
Computes the id range for a single fn body, ignoring nested items.

compute_id_range_for_inlined_item 	 [Unstable]

empty_generics 	 [Unstable]

float_ty_to_string 	 [Unstable]

ident_to_pat 	 [Unstable]

ident_to_path 	 [Unstable]

impl_pretty_name 	 [Unstable]
Generate a "pretty" name for an impl from its type and trait. This is designed so that symbols of impl'd methods give some hint of where they came from, (previously they would all just be listed as __extensions__::method_name::hash, with no indication of the type).

int_ty_max 	 [Unstable]

int_ty_to_string 	 [Unstable]
Get a string representation of a signed int type, with its value. We want to avoid "45int" and "-3int" in favor of "45" and "-3"

is_by_value_binop 	 [Unstable]
Returns true if the binary operator takes its arguments by value

is_by_value_unop 	 [Unstable]
Returns true if the unary operator takes its argument by value

is_comparison_binop 	 [Unstable]

is_local 	 [Unstable]

is_path 	 [Unstable]

is_shift_binop 	 [Unstable]

lazy_binop 	 [Unstable]

lit_is_str 	 [Unstable]
Returns true if this literal is a string and false otherwise.

local_def 	 [Unstable]

name_to_dummy_lifetime 	 [Unstable]

operator_prec 	 [Unstable]
Maps a binary operator to its precedence

pat_is_ident 	 [Unstable]
Returns true if the given pattern consists solely of an identifier and false otherwise.

path_name_eq 	 [Unstable]

path_name_i 	 [Unstable]

path_to_ident 	 [Unstable]

segments_name_eq 	 [Unstable]

stmt_id 	 [Unstable]

struct_def_is_tuple_like 	 [Unstable]
Returns true if the given struct def is tuple-like; i.e. that its fields are unnamed.

struct_field_visibility 	 [Unstable]

uint_ty_max 	 [Unstable]

uint_ty_to_string 	 [Unstable]
Get a string representation of an unsigned int type, with its value. We want to avoid "42u" in favor of "42us". "42uint" is right out.

unop_to_string 	 [Unstable]

visit_ids_for_inlined_item 	 [Unstable]

walk_pat 	 [Unstable]