Module syntax::attr

Functions dealing with attributes and meta items

Re-exports

pub use IntType::*;

pub use ReprAttr::*;

pub use StabilityLevel::*;

Structs

Deprecation

RustcDeprecation

Stability
Represents the #[stable], #[unstable], #[rustc_{deprecated,const_unstable}] attributes.

Enums

InlineAttr

IntType

OptimizeAttr

ReprAttr

StabilityLevel
The available stability levels.

UnwindAttr

Traits

HasAttrs

Functions

cfg_matches
Tests if a cfg-pattern matches the cfg set

contains_feature_attr
Checks if attrs contains an attribute like #![feature(feature_name)]. This will not perform any "sanity checks" on the form of the attributes.

contains_name

eval_condition
Evaluate a cfg-like condition (with any and all), using eval to evaluate individual items.

filter_by_name

find_by_name

find_crate_name

find_deprecation
Finds the deprecation attribute. None if none exists.

find_repr_attrs
Parse #[repr(...)] forms.

find_stability
Finds the first stability attribute. None if none exists.

find_unwind_attr
Determine what #[unwind] attribute is present in attrs, if any.

first_attr_value_str_by_name

inject

is_known

is_known_lint_tool

is_used

list_contains_name

mark_known

mark_used

mk_attr_id

mk_attr_inner
Returns an inner attribute with the given value.

mk_attr_outer
Returns an outer attribute with the given value.

mk_list_item

mk_name_value_item

mk_name_value_item_str

mk_nested_word_item

mk_spanned_attr_inner
Returns an inner attribute with the given value and span.

mk_spanned_attr_outer
Returns an outer attribute with the given value and span.

mk_sugared_doc_attr

mk_word_item