Trait syntax::attr::HasAttrs

pub trait HasAttrs: Sized {
    fn attrs(&self) -> &[Attribute];

    fn map_attrs<F: FnOnce(Vec<Attribute>) -> Vec<Attribute>>(
        self, 
        f: F
    ) -> Self;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn attrs(&self) -> &[Attribute]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn map_attrs<F: FnOnce(Vec<Attribute>) -> Vec<Attribute>>(self, f: F) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?