Struct syntax::codemap::SpanUnstable

pub struct Span {
    pub lo: BytePos,
    pub hi: BytePos,
    pub expn_id: ExpnId,
}

Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source.

Fields

lo

hi

expn_id

Information about where the macro came from, if this piece of code was created by a macro expansion.