Trait syntax::codemap::FileLoader

pub trait FileLoader {
    fn file_exists(&self, path: &Path) -> bool;

    fn abs_path(&self, path: &Path) -> Option<PathBuf>;

    fn read_file(&self, path: &Path) -> Result<String>;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An abstraction over the fs operations used by the Parser.

Required Methods

fn file_exists(&self, path: &Path) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Query the existence of a file.

fn abs_path(&self, path: &Path) -> Option<PathBuf>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Return an absolute path to a file, if possible.

fn read_file(&self, path: &Path) -> Result<String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Read the contents of an UTF-8 file into memory.