Module syntax::cst

Structs

Case

Cocase

Codata

Codef

Comatch

Ctor

Data

Def

Destructee

Dtor

Info

Match

Motive

Param

A Param can either be a single parameter, like x : T, or a list of parameters, like x, y, z : T.

ParamInst

Instantiation of a previously declared parameter

Prg

Scrutinee

SelfParam

Telescope

Wrapper type signifying the wrapped parameters have telescope semantics. I.e. each parameter binding in the parameter list is in scope for the following parameters. This influences the lowering semantic.

TelescopeInst

Instantiation of a previously declared telescope

TypApp

Enums

DefDecl

Exp

Item

TypDecl

Type Definitions

Args

Params