Trait syntax::ctx::BindContext

pub trait BindContext: Sized {
    type Ctx: Context;

    fn ctx_mut(&mut self) -> &mut Self::Ctx;

    fn bind_single<T, O, F>(&mut self, elem: T, f: F) -> O
    where
        T: ContextElem<Self::Ctx>,
        F: FnOnce(&mut Self) -> O,
    { ... }
    fn bind_iter<T, I, O, F>(&mut self, iter: I, f: F) -> O
    where
        T: ContextElem<Self::Ctx>,
        I: Iterator<Item = T>,
        F: FnOnce(&mut Self) -> O,
    { ... }
    fn bind_fold<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
        &mut self,
        iter: I,
        acc: O1,
        f_acc: F1,
        f_inner: F2
    ) -> O2
    where
        T: ContextElem<Self::Ctx>,
        F1: Fn(&mut Self, O1, T) -> O1,
        F2: FnOnce(&mut Self, O1) -> O2,
    { ... }
    fn bind_fold2<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
        &mut self,
        iter: I,
        acc: O1,
        f_acc: F1,
        f_inner: F2
    ) -> O2
    where
        F1: Fn(&mut Self, O1, T) -> BindElem<<Self::Ctx as Context>::ElemIn, O1>,
        F2: FnOnce(&mut Self, O1) -> O2,
    { ... }
    fn bind_fold_failable<T, I: Iterator<Item = T>, O1, O2, F1, F2, E>(
        &mut self,
        iter: I,
        acc: O1,
        f_acc: F1,
        f_inner: F2
    ) -> Result<O2, E>
    where
        F1: Fn(&mut Self, O1, T) -> Result<BindElem<<Self::Ctx as Context>::ElemIn, O1>, E>,
        F2: FnOnce(&mut Self, O1) -> O2,
    { ... }
}

Interface to bind variables to anything that has a Context

There are two use cases for this trait.

Case 1: You have a type that implements Context. Then, a blanket impl ensures that this type also implements BindContext.

Case 2: You have a type that has a field which implements Context. Then, only implement the ctx_mut method for BindContext and return the field that implements Context. Do not override the default implementations for the bind_* methods.

In both cases, BindContext will provide a safe interface to bind variables and telescopes.

Required Associated Types

type Ctx: Context

Required Methods

fn ctx_mut(&mut self) -> &mut Self::Ctx

Provided Methods

fn bind_single<T, O, F>(&mut self, elem: T, f: F) -> Owhere
T: ContextElem<Self::Ctx>,
F: FnOnce(&mut Self) -> O,

Bind a single element

fn bind_iter<T, I, O, F>(&mut self, iter: I, f: F) -> Owhere
T: ContextElem<Self::Ctx>,
I: Iterator<Item = T>,
F: FnOnce(&mut Self) -> O,

Bind an iterator iter of binders

fn bind_fold<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> O2where
T: ContextElem<Self::Ctx>,
F1: Fn(&mut Self, O1, T) -> O1,
F2: FnOnce(&mut Self, O1) -> O2,

Bind an iterator iter of elements

Fold the iterator and consume the result under the inner context with all binders in scope.

This is used for checking telescopes.

iter - An iterator of binders

acc - Accumulator for folding the iterator

f_acc - Accumulator function run for each binder

f_inner - Inner function computing the final result under the context of all binders

fn bind_fold2<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> O2where
F1: Fn(&mut Self, O1, T) -> BindElem<<Self::Ctx as Context>::ElemIn, O1>,
F2: FnOnce(&mut Self, O1) -> O2,

fn bind_fold_failable<T, I: Iterator<Item = T>, O1, O2, F1, F2, E>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> Result<O2, E>where
F1: Fn(&mut Self, O1, T) -> Result<BindElem<<Self::Ctx as Context>::ElemIn, O1>, E>,
F2: FnOnce(&mut Self, O1) -> O2,