Trait syntax::ctx::Context

pub trait Context: Sized {
    type ElemIn: Clone;
    type ElemOut;
    type Var;

    fn empty() -> Self;
    fn lookup<V: Into<Self::Var> + Debug>(&self, var: V) -> Self::ElemOut;
    fn push_telescope(&mut self);
    fn pop_telescope(&mut self);
    fn push_binder(&mut self, elem: Self::ElemIn);
    fn pop_binder(&mut self, elem: Self::ElemIn);
}

Defines the interface of a variable context

Required Associated Types

type ElemIn: Clone

The type of elements that can be bound to the context

type ElemOut

Result type of a lookup into the context

type Var

Variable type that can be looked up in the context

Required Methods

fn empty() -> Self

Create an empty context

fn lookup<V: Into<Self::Var> + Debug>(&self, var: V) -> Self::ElemOut

Get the element bound at var

fn push_telescope(&mut self)

Add a new telescope to the context This function is run if a new list of binders should be pushed to the context

fn pop_telescope(&mut self)

Remove a telescope from the context Function that is run if the most-recently pushed list of binders should be unbound from the context

fn push_binder(&mut self, elem: Self::ElemIn)

Push a binder into the most-recently pushed telescope

fn pop_binder(&mut self, elem: Self::ElemIn)

Pop a binder from the most-recently pushed telescope