Trait syntax::de_bruijn::Leveled

pub trait Leveled {
    fn idx_to_lvl(&self, idx: Idx) -> Lvl;
    fn lvl_to_idx(&self, lvl: Lvl) -> Idx;

    fn var_to_lvl(&self, var: Var) -> Lvl { ... }
    fn var_to_idx(&self, var: Var) -> Idx { ... }
}

Convert an De-Bruijn index to a De-Bruijn level

To perform this conversion, it is sufficient to track

the current number of De-Bruijn levels (maximum first component)

the current number of binders per De-Bruijn level (maximum second component for each first component). This is satisfied by the context type during typechecking.

Required Methods

fn idx_to_lvl(&self, idx: Idx) -> Lvl

fn lvl_to_idx(&self, lvl: Lvl) -> Idx

Provided Methods

fn var_to_lvl(&self, var: Var) -> Lvl

fn var_to_idx(&self, var: Var) -> Idx