Struct syntax::diagnostic::SpanHandler

pub struct SpanHandler {
    pub handler: Handler,
    pub cm: CodeMap,
}

Unstable (rustc_private)

A span-handler is like a handler but also accepts span information for source-location reporting.

Fields

handler
Unstable (rustc_private)

cm
Unstable (rustc_private)

Methods

impl SpanHandler

fn new(handler: Handler, cm: CodeMap) -> SpanHandler

Unstable (rustc_private)

fn span_fatal(&self, sp: Span, msg: &str) -> FatalError

Unstable (rustc_private)

fn span_fatal_with_code(&self, sp: Span, msg: &str, code: &str) -> FatalError

Unstable (rustc_private)

fn span_err(&self, sp: Span, msg: &str)

Unstable (rustc_private)

fn span_err_with_code(&self, sp: Span, msg: &str, code: &str)

Unstable (rustc_private)

fn span_warn(&self, sp: Span, msg: &str)

Unstable (rustc_private)

fn span_warn_with_code(&self, sp: Span, msg: &str, code: &str)

Unstable (rustc_private)

fn span_note(&self, sp: Span, msg: &str)

Unstable (rustc_private)

fn span_end_note(&self, sp: Span, msg: &str)

Unstable (rustc_private)

fn span_help(&self, sp: Span, msg: &str)

Unstable (rustc_private)

fn span_suggestion(&self, sp: Span, msg: &str, suggestion: String)

Unstable (rustc_private)

Prints out a message with a suggested edit of the code.

See diagnostic::RenderSpan::Suggestion for more information.

fn fileline_note(&self, sp: Span, msg: &str)

Unstable (rustc_private)

fn fileline_help(&self, sp: Span, msg: &str)

Unstable (rustc_private)

fn span_bug(&self, sp: Span, msg: &str) -> !

Unstable (rustc_private)

fn span_unimpl(&self, sp: Span, msg: &str) -> !

Unstable (rustc_private)

fn handler<'a>(&'a self) -> &'a Handler

Unstable (rustc_private)