Struct syntax::diagnostics::metadata::ErrorMetadata

pub struct ErrorMetadata {
    pub description: Option<String>,
    pub use_site: Option<ErrorLocation>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

JSON encodable/decodable version of ErrorInfo.

Fields
description: Option<String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

use_site: Option<ErrorLocation>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?