Struct syntax::diagnostics::plugin::DiagnosticBuilder

#[must_use]

pub struct DiagnosticBuilder<'a> {
    pub handler: &'a Handler,
    // some fields omitted
}

Used for emitting structured error messages and other diagnostic information.

If there is some state in a downstream crate you would like to access in the methods of DiagnosticBuilder here, consider extending HandlerFlags, accessed via self.handler.flags.

Fields
handler: &'a Handler
Methods

impl<'a> DiagnosticBuilder<'a>

pub fn emit(&mut self)

Emit the diagnostic.

pub fn buffer(self, buffered_diagnostics: &mut Vec<Diagnostic>)

Buffers the diagnostic for later emission, unless handler has disabled such buffering.

pub fn sub<S>(
&mut self,
level: Level,
message: &str,
span: Option<S>
) -> &mut DiagnosticBuilder<'a> where
S: Into<MultiSpan>,

Convenience function for internal use, clients should use one of the span_* methods instead.

pub fn delay_as_bug(&mut self)

Delay emission of this diagnostic as a bug.

This can be useful in contexts where an error indicates a bug but typically this only happens when other compilation errors have already happened. In those cases this can be used to defer emission of this diagnostic as a bug in the compiler only if no other errors have been emitted.

In the meantime, though, callsites are required to deal with the "bug" locally in whichever way makes the most sense.

pub fn span_label<T>(
&mut self,
span: Span,
label: T
) -> &mut DiagnosticBuilder<'a> where
T: Into<String>,

Adds a span/label to be included in the resulting snippet. This is pushed onto the MultiSpan that was created when the diagnostic was first built. If you don't call this function at all, and you just supplied a Span to create the diagnostic, then the snippet will just include that Span, which is called the primary span.

pub fn note_expected_found(
&mut self,
label: &dyn Display,
expected: DiagnosticStyledString,
found: DiagnosticStyledString
) -> &mut DiagnosticBuilder<'a>

pub fn note_expected_found_extra(
&mut self,
label: &dyn Display,
expected: DiagnosticStyledString,
found: DiagnosticStyledString,
expected_extra: &dyn Display,
found_extra: &dyn Display
) -> &mut DiagnosticBuilder<'a>

pub fn note(&mut self, msg: &str) -> &mut DiagnosticBuilder<'a>

pub fn span_note<S>(&mut self, sp: S, msg: &str) -> &mut DiagnosticBuilder<'a> where
S: Into<MultiSpan>,

pub fn warn(&mut self, msg: &str) -> &mut DiagnosticBuilder<'a>

pub fn span_warn<S>(&mut self, sp: S, msg: &str) -> &mut DiagnosticBuilder<'a> where
S: Into<MultiSpan>,

pub fn help(&mut self, msg: &str) -> &mut DiagnosticBuilder<'a>

pub fn span_help<S>(&mut self, sp: S, msg: &str) -> &mut DiagnosticBuilder<'a> where
S: Into<MultiSpan>,

pub fn multipart_suggestion(
&mut self,
msg: &str,
suggestion: Vec<(Span, String)>,
applicability: Applicability
) -> &mut DiagnosticBuilder<'a>

pub fn tool_only_multipart_suggestion(
&mut self,
msg: &str,
suggestion: Vec<(Span, String)>,
applicability: Applicability
) -> &mut DiagnosticBuilder<'a>

pub fn span_suggestion(
&mut self,
sp: Span,
msg: &str,
suggestion: String,
applicability: Applicability
) -> &mut DiagnosticBuilder<'a>

pub fn span_suggestions<impl Iterator>(
&mut self,
sp: Span,
msg: &str,
suggestions: impl Iterator,
applicability: Applicability
) -> &mut DiagnosticBuilder<'a> where
impl Iterator: Iterator<Item = String>,

pub fn span_suggestion_short(
&mut self,
sp: Span,
msg: &str,
suggestion: String,
applicability: Applicability
) -> &mut DiagnosticBuilder<'a>

pub fn span_suggestion_hidden(
&mut self,
sp: Span,
msg: &str,
suggestion: String,
applicability: Applicability
) -> &mut DiagnosticBuilder<'a>

pub fn tool_only_span_suggestion(
&mut self,
sp: Span,
msg: &str,
suggestion: String,
applicability: Applicability
) -> &mut DiagnosticBuilder<'a>

pub fn set_span<S>(&mut self, sp: S) -> &mut DiagnosticBuilder<'a> where
S: Into<MultiSpan>,

pub fn code(&mut self, s: DiagnosticId) -> &mut DiagnosticBuilder<'a>

pub fn allow_suggestions(&mut self, allow: bool) -> &mut DiagnosticBuilder<'a>

pub fn new(
handler: &'a Handler,
level: Level,
message: &str
) -> DiagnosticBuilder<'a>

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub fn new_with_code(
handler: &'a Handler,
level: Level,
code: Option<DiagnosticId>,
message: &str
) -> DiagnosticBuilder<'a>

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub fn new_diagnostic(
handler: &'a Handler,
diagnostic: Diagnostic
) -> DiagnosticBuilder<'a>

Creates a new DiagnosticBuilder with an already constructed diagnostic.