Enum syntax::NodeOrToken

pub enum NodeOrToken<N, T> {
    Node(N),
    Token(T),
}

Variants

Node(N)

Token(T)

Implementations

impl<'_> NodeOrToken<&'_ GreenNodeData, &'_ GreenTokenData>

pub fn to_owned(self) -> NodeOrToken<GreenNode, GreenToken>

impl NodeOrToken<GreenNode, GreenToken>

pub fn kind(&self) -> SyntaxKind

Returns kind of this element.

pub fn text_len(&self) -> TextSize

Returns the length of the text covered by this element.

impl<'_> NodeOrToken<&'_ GreenNodeData, &'_ GreenTokenData>

pub fn kind(&self) -> SyntaxKind

Returns kind of this element.

pub fn text_len(self) -> TextSize

Returns the length of the text covered by this element.

impl NodeOrToken<SyntaxNode, SyntaxToken>

pub fn text_range(&self) -> TextRange

pub fn index(&self) -> usize

pub fn kind(&self) -> SyntaxKind

pub fn parent(&self) -> Option<SyntaxNode>

pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode>

pub fn first_token(&self) -> Option<SyntaxToken>

pub fn last_token(&self) -> Option<SyntaxToken>

pub fn next_sibling_or_token(
&self
) -> Option<NodeOrToken<SyntaxNode, SyntaxToken>>

pub fn prev_sibling_or_token(
&self
) -> Option<NodeOrToken<SyntaxNode, SyntaxToken>>

pub fn detach(&self)

impl<L> NodeOrToken<SyntaxNode<L>, SyntaxToken<L>> where
L: Language,

pub fn text_range(&self) -> TextRange

pub fn index(&self) -> usize

pub fn kind(&self) -> <L as Language>::Kind

pub fn parent(&self) -> Option<SyntaxNode<L>>

pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode<L>>

pub fn next_sibling_or_token(
&self
) -> Option<NodeOrToken<SyntaxNode<L>, SyntaxToken<L>>>

pub fn prev_sibling_or_token(
&self
) -> Option<NodeOrToken<SyntaxNode<L>, SyntaxToken<L>>>

pub fn detach(&self)

impl<N, T> NodeOrToken<N, T>

pub fn into_node(self) -> Option<N>

pub fn into_token(self) -> Option<T>

pub fn as_node(&self) -> Option<&N>

pub fn as_token(&self) -> Option<&T>