syntax

Enum SyntaxKind

pub enum SyntaxKind {

    DOLLAR,
    SEMICOLON,
    COMMA,
    L_PAREN,
    R_PAREN,
    L_CURLY,
    R_CURLY,
    L_BRACK,
    R_BRACK,
    L_ANGLE,
    R_ANGLE,
    AT,
    POUND,
    TILDE,
    QUESTION,
    AMP,
    PIPE,
    PLUS,
    STAR,
    SLASH,
    CARET,
    PERCENT,
    UNDERSCORE,
    DOT,
    DOT2,
    DOT3,
    DOT2EQ,
    COLON,
    COLON2,
    EQ,
    EQ2,
    FAT_ARROW,
    BANG,
    NEQ,
    MINUS,
    THIN_ARROW,
    LTEQ,
    GTEQ,
    PLUSEQ,
    MINUSEQ,
    PIPEEQ,
    AMPEQ,
    CARETEQ,
    SLASHEQ,
    STAREQ,
    PERCENTEQ,
    AMP2,
    PIPE2,
    SHL,
    SHR,
    SHLEQ,
    SHREQ,
    SELF_TYPE_KW,
    ABSTRACT_KW,
    AS_KW,
    BECOME_KW,
    BOX_KW,
    BREAK_KW,
    CONST_KW,
    CONTINUE_KW,
    CRATE_KW,
    DO_KW,
    ELSE_KW,
    ENUM_KW,
    EXTERN_KW,
    FALSE_KW,
    FINAL_KW,
    FN_KW,
    FOR_KW,
    IF_KW,
    IMPL_KW,
    IN_KW,
    LET_KW,
    LOOP_KW,
    MACRO_KW,
    MATCH_KW,
    MOD_KW,
    MOVE_KW,
    MUT_KW,
    OVERRIDE_KW,
    PRIV_KW,
    PUB_KW,
    REF_KW,
    RETURN_KW,
    SELF_KW,
    STATIC_KW,
    STRUCT_KW,
    SUPER_KW,
    TRAIT_KW,
    TRUE_KW,
    TYPE_KW,
    TYPEOF_KW,
    UNSAFE_KW,
    UNSIZED_KW,
    USE_KW,
    VIRTUAL_KW,
    WHERE_KW,
    WHILE_KW,
    YIELD_KW,
    ASM_KW,
    ASYNC_KW,
    ATT_SYNTAX_KW,
    AUTO_KW,
    AWAIT_KW,
    BUILTIN_KW,
    CLOBBER_ABI_KW,
    DEFAULT_KW,
    DYN_KW,
    FORMAT_ARGS_KW,
    GEN_KW,
    GLOBAL_ASM_KW,
    INLATEOUT_KW,
    INOUT_KW,
    LABEL_KW,
    LATEOUT_KW,
    MACRO_RULES_KW,
    MAY_UNWIND_KW,
    NAKED_ASM_KW,
    NOMEM_KW,
    NORETURN_KW,
    NOSTACK_KW,
    OFFSET_OF_KW,
    OPTIONS_KW,
    OUT_KW,
    PRESERVES_FLAGS_KW,
    PURE_KW,
    RAW_KW,
    READONLY_KW,
    SAFE_KW,
    SYM_KW,
    TRY_KW,
    UNION_KW,
    YEET_KW,
    BYTE,
    BYTE_STRING,
    CHAR,
    C_STRING,
    FLOAT_NUMBER,
    INT_NUMBER,
    STRING,
    COMMENT,
    ERROR,
    FRONTMATTER,
    IDENT,
    LIFETIME_IDENT,
    NEWLINE,
    SHEBANG,
    WHITESPACE,
    ABI,
    ARG_LIST,
    ARRAY_EXPR,
    ARRAY_TYPE,
    ASM_CLOBBER_ABI,
    ASM_CONST,
    ASM_DIR_SPEC,
    ASM_EXPR,
    ASM_LABEL,
    ASM_OPERAND_EXPR,
    ASM_OPERAND_NAMED,
    ASM_OPTION,
    ASM_OPTIONS,
    ASM_REG_OPERAND,
    ASM_REG_SPEC,
    ASM_SYM,
    ASSOC_ITEM_LIST,
    ASSOC_TYPE_ARG,
    ATTR,
    AWAIT_EXPR,
    BECOME_EXPR,
    BIN_EXPR,
    BLOCK_EXPR,
    BOX_PAT,
    BREAK_EXPR,
    CALL_EXPR,
    CAST_EXPR,
    CLOSURE_EXPR,
    CONST,
    CONST_ARG,
    CONST_BLOCK_PAT,
    CONST_PARAM,
    CONTINUE_EXPR,
    DYN_TRAIT_TYPE,
    ENUM,
    EXPR_STMT,
    EXTERN_BLOCK,
    EXTERN_CRATE,
    EXTERN_ITEM_LIST,
    FIELD_EXPR,
    FN,
    FN_PTR_TYPE,
    FORMAT_ARGS_ARG,
    FORMAT_ARGS_ARG_NAME,
    FORMAT_ARGS_EXPR,
    FOR_BINDER,
    FOR_EXPR,
    FOR_TYPE,
    GENERIC_ARG_LIST,
    GENERIC_PARAM_LIST,
    IDENT_PAT,
    IF_EXPR,
    IMPL,
    IMPL_TRAIT_TYPE,
    INDEX_EXPR,
    INFER_TYPE,
    ITEM_LIST,
    LABEL,
    LET_ELSE,
    LET_EXPR,
    LET_STMT,
    LIFETIME,
    LIFETIME_ARG,
    LIFETIME_PARAM,
    LITERAL,
    LITERAL_PAT,
    LOOP_EXPR,
    MACRO_CALL,
    MACRO_DEF,
    MACRO_EXPR,
    MACRO_ITEMS,
    MACRO_PAT,
    MACRO_RULES,
    MACRO_STMTS,
    MACRO_TYPE,
    MATCH_ARM,
    MATCH_ARM_LIST,
    MATCH_EXPR,
    MATCH_GUARD,
    META,
    METHOD_CALL_EXPR,
    MODULE,
    NAME,
    NAME_REF,
    NEVER_TYPE,
    OFFSET_OF_EXPR,
    OR_PAT,
    PARAM,
    PARAM_LIST,
    PARENTHESIZED_ARG_LIST,
    PAREN_EXPR,
    PAREN_PAT,
    PAREN_TYPE,
    PATH,
    PATH_EXPR,
    PATH_PAT,
    PATH_SEGMENT,
    PATH_TYPE,
    PREFIX_EXPR,
    PTR_TYPE,
    RANGE_EXPR,
    RANGE_PAT,
    RECORD_EXPR,
    RECORD_EXPR_FIELD,
    RECORD_EXPR_FIELD_LIST,
    RECORD_FIELD,
    RECORD_FIELD_LIST,
    RECORD_PAT,
    RECORD_PAT_FIELD,
    RECORD_PAT_FIELD_LIST,
    REF_EXPR,
    REF_PAT,
    REF_TYPE,
    RENAME,
    REST_PAT,
    RETURN_EXPR,
    RETURN_TYPE_SYNTAX,
    RET_TYPE,
    SELF_PARAM,
    SLICE_PAT,
    SLICE_TYPE,
    SOURCE_FILE,
    STATIC,
    STMT_LIST,
    STRUCT,
    TOKEN_TREE,
    TRAIT,
    TRY_EXPR,
    TUPLE_EXPR,
    TUPLE_FIELD,
    TUPLE_FIELD_LIST,
    TUPLE_PAT,
    TUPLE_STRUCT_PAT,
    TUPLE_TYPE,
    TYPE_ALIAS,
    TYPE_ANCHOR,
    TYPE_ARG,
    TYPE_BOUND,
    TYPE_BOUND_LIST,
    TYPE_PARAM,
    UNDERSCORE_EXPR,
    UNION,
    USE,
    USE_BOUND_GENERIC_ARGS,
    USE_TREE,
    USE_TREE_LIST,
    VARIANT,
    VARIANT_LIST,
    VISIBILITY,
    WHERE_CLAUSE,
    WHERE_PRED,
    WHILE_EXPR,
    WILDCARD_PAT,
    YEET_EXPR,
    YIELD_EXPR,
    // some variants omitted

}

The kind of syntax node, e.g. IDENT, USE_KW, or STRUCT.

Variants

DOLLAR

SEMICOLON

COMMA

L_PAREN

R_PAREN

L_CURLY

R_CURLY

L_BRACK

R_BRACK

L_ANGLE

R_ANGLE

AT

POUND

TILDE

QUESTION

AMP

PIPE

PLUS

STAR

SLASH

CARET

PERCENT

UNDERSCORE

DOT

DOT2

DOT3

DOT2EQ

COLON

COLON2

EQ

EQ2

FAT_ARROW

BANG

NEQ

MINUS

THIN_ARROW

LTEQ

GTEQ

PLUSEQ

MINUSEQ

PIPEEQ

AMPEQ

CARETEQ

SLASHEQ

STAREQ

PERCENTEQ

AMP2

PIPE2

SHL

SHR

SHLEQ

SHREQ

SELF_TYPE_KW

ABSTRACT_KW

AS_KW

BECOME_KW

BOX_KW

BREAK_KW

CONST_KW

CONTINUE_KW

CRATE_KW

DO_KW

ELSE_KW

ENUM_KW

EXTERN_KW

FALSE_KW

FINAL_KW

FN_KW

FOR_KW

IF_KW

IMPL_KW

IN_KW

LET_KW

LOOP_KW

MACRO_KW

MATCH_KW

MOD_KW

MOVE_KW

MUT_KW

OVERRIDE_KW

PRIV_KW

PUB_KW

REF_KW

RETURN_KW

SELF_KW

STATIC_KW

STRUCT_KW

SUPER_KW

TRAIT_KW

TRUE_KW

TYPE_KW

TYPEOF_KW

UNSAFE_KW

UNSIZED_KW

USE_KW

VIRTUAL_KW

WHERE_KW

WHILE_KW

YIELD_KW

ASM_KW

ASYNC_KW

ATT_SYNTAX_KW

AUTO_KW

AWAIT_KW

BUILTIN_KW

CLOBBER_ABI_KW

DEFAULT_KW

DYN_KW

FORMAT_ARGS_KW

GEN_KW

GLOBAL_ASM_KW

INLATEOUT_KW

INOUT_KW

LABEL_KW

LATEOUT_KW

MACRO_RULES_KW

MAY_UNWIND_KW

NAKED_ASM_KW

NOMEM_KW

NORETURN_KW

NOSTACK_KW

OFFSET_OF_KW

OPTIONS_KW

OUT_KW

PRESERVES_FLAGS_KW

PURE_KW

RAW_KW

READONLY_KW

SAFE_KW

SYM_KW

TRY_KW

UNION_KW

YEET_KW

BYTE

BYTE_STRING

CHAR

C_STRING

FLOAT_NUMBER

INT_NUMBER

STRING

COMMENT

ERROR

FRONTMATTER

IDENT

LIFETIME_IDENT

NEWLINE

SHEBANG

WHITESPACE

ABI

ARG_LIST

ARRAY_EXPR

ARRAY_TYPE

ASM_CLOBBER_ABI

ASM_CONST

ASM_DIR_SPEC

ASM_EXPR

ASM_LABEL

ASM_OPERAND_EXPR

ASM_OPERAND_NAMED

ASM_OPTION

ASM_OPTIONS

ASM_REG_OPERAND

ASM_REG_SPEC

ASM_SYM

ASSOC_ITEM_LIST

ASSOC_TYPE_ARG

ATTR

AWAIT_EXPR

BECOME_EXPR

BIN_EXPR

BLOCK_EXPR

BOX_PAT

BREAK_EXPR

CALL_EXPR

CAST_EXPR

CLOSURE_EXPR

CONST

CONST_ARG

CONST_BLOCK_PAT

CONST_PARAM

CONTINUE_EXPR

DYN_TRAIT_TYPE

ENUM

EXPR_STMT

EXTERN_BLOCK

EXTERN_CRATE

EXTERN_ITEM_LIST

FIELD_EXPR

FN

FN_PTR_TYPE

FORMAT_ARGS_ARG

FORMAT_ARGS_ARG_NAME

FORMAT_ARGS_EXPR

FOR_BINDER

FOR_EXPR

FOR_TYPE

GENERIC_ARG_LIST

GENERIC_PARAM_LIST

IDENT_PAT

IF_EXPR

IMPL

IMPL_TRAIT_TYPE

INDEX_EXPR

INFER_TYPE

ITEM_LIST

LABEL

LET_ELSE

LET_EXPR

LET_STMT

LIFETIME

LIFETIME_ARG

LIFETIME_PARAM

LITERAL

LITERAL_PAT

LOOP_EXPR

MACRO_CALL

MACRO_DEF

MACRO_EXPR

MACRO_ITEMS

MACRO_PAT

MACRO_RULES

MACRO_STMTS

MACRO_TYPE

MATCH_ARM

MATCH_ARM_LIST

MATCH_EXPR

MATCH_GUARD

META

METHOD_CALL_EXPR

MODULE

NAME

NAME_REF

NEVER_TYPE

OFFSET_OF_EXPR

OR_PAT

PARAM

PARAM_LIST

PARENTHESIZED_ARG_LIST

PAREN_EXPR

PAREN_PAT

PAREN_TYPE

PATH

PATH_EXPR

PATH_PAT

PATH_SEGMENT

PATH_TYPE

PREFIX_EXPR

PTR_TYPE

RANGE_EXPR

RANGE_PAT

RECORD_EXPR

RECORD_EXPR_FIELD

RECORD_EXPR_FIELD_LIST

RECORD_FIELD

RECORD_FIELD_LIST

RECORD_PAT

RECORD_PAT_FIELD

RECORD_PAT_FIELD_LIST

REF_EXPR

REF_PAT

REF_TYPE

RENAME

REST_PAT

RETURN_EXPR

RETURN_TYPE_SYNTAX

RET_TYPE

SELF_PARAM

SLICE_PAT

SLICE_TYPE

SOURCE_FILE

STATIC

STMT_LIST

STRUCT

TOKEN_TREE

TRAIT

TRY_EXPR

TUPLE_EXPR

TUPLE_FIELD

TUPLE_FIELD_LIST

TUPLE_PAT

TUPLE_STRUCT_PAT

TUPLE_TYPE

TYPE_ALIAS

TYPE_ANCHOR

TYPE_ARG

TYPE_BOUND

TYPE_BOUND_LIST

TYPE_PARAM

UNDERSCORE_EXPR

UNION

USE

USE_BOUND_GENERIC_ARGS

USE_TREE

USE_TREE_LIST

VARIANT

VARIANT_LIST

VISIBILITY

WHERE_CLAUSE

WHERE_PRED

WHILE_EXPR

WILDCARD_PAT

YEET_EXPR

YIELD_EXPR

Implementations

impl SyntaxKind

pub const fn text(self) -> &'static str

pub fn is_strict_keyword(self, edition: Edition) -> bool

Checks whether this syntax kind is a strict keyword for the given edition. Strict keywords are identifiers that are always considered keywords.

pub fn is_contextual_keyword(self, edition: Edition) -> bool

Checks whether this syntax kind is a weak keyword for the given edition. Weak keywords are identifiers that are considered keywords only in certain contexts.

pub fn is_keyword(self, edition: Edition) -> bool

Checks whether this syntax kind is a strict or weak keyword for the given edition.

pub fn is_punct(self) -> bool

pub fn is_literal(self) -> bool

pub fn from_keyword(ident: &str, edition: Edition) -> Option<SyntaxKind>

pub fn from_contextual_keyword( ident: &str, edition: Edition, ) -> Option<SyntaxKind>

pub fn from_char(c: char) -> Option<SyntaxKind>

impl SyntaxKind

pub fn is_trivia(self) -> bool

pub fn is_any_identifier(self) -> bool

Returns true if this is an identifier or a keyword.