Module syntax::errors

Unstable (rustc_private)

Reexports

pub use errors::emitter::ColorConfig;

Modules

emitter 	 [Unstable]

json 	 [Unstable]
A JSON emitter for errors.

snippet 	 [Unstable]

Structs

CodeSuggestion 	 [Unstable]

DiagnosticBuilder 	 [Unstable]
Used for emitting structured error messages and other diagnostic information.

ExplicitBug 	 [Unstable]
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError 	 [Unstable]
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

Handler 	 [Unstable]
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

Enums

Level 	 [Unstable]

RenderSpan 	 [Unstable]

Functions

check_old_skool 	 [Unstable]
True if we should use the old-skool error format style. This is the default setting until the new errors are deemed stable enough for general use.

expect 	 [Unstable]