Enum syntax::ext::base::SyntaxExtension

pub enum SyntaxExtension {
    Decorator(Box<ItemDecorator + 'static>),
    MultiDecorator(Box<MultiItemDecorator + 'static>),
    Modifier(Box<ItemModifier + 'static>),
    MultiModifier(Box<MultiItemModifier + 'static>),
    NormalTT(Box<TTMacroExpander + 'static>, Option<Span>, bool),
    IdentTT(Box<IdentMacroExpander + 'static>, Option<Span>, bool),
    MacroRulesTT,
}

Unstable

An enum representing the different kinds of syntax extensions.

Variants

Decorator
Deprecated since 1.0.0
: replaced by MultiDecorator

A syntax extension that is attached to an item and creates new items based upon it.

MultiDecorator
Unstable

A syntax extension that is attached to an item and creates new items based upon it.

#[derive(...)] is a MultiItemDecorator.

Modifier
Deprecated since 1.0.0
: replaced by MultiModifier

A syntax extension that is attached to an item and modifies it in-place.

MultiModifier
Unstable

A syntax extension that is attached to an item and modifies it in-place. More flexible version than Modifier.

NormalTT
Unstable

A normal, function-like syntax extension.

bytes! is a NormalTT.

The bool dictates whether the contents of the macro can directly use #[unstable] things (true == yes).

IdentTT
Unstable

A function-like syntax extension that has an extra ident before the block.

MacroRulesTT
Unstable

Represents macro_rules! itself.