Struct syntax::ext::base::ExtCtxt

pub struct ExtCtxt<'a> {
    pub parse_sess: &'a ParseSess,
    pub ecfg: ExpansionConfig<'a>,
    pub root_path: PathBuf,
    pub resolver: &'a mut dyn Resolver,
    pub current_expansion: ExpansionData,
    pub expansions: FxHashMap<Span, Vec<String>>,
}

One of these is made during expansion and incrementally updated as we go; when a macro expansion occurs, the resulting nodes have the backtrace() -> expn_info of their expansion context stored into their span.

Fields
parse_sess: &'a ParseSessecfg: ExpansionConfig<'a>root_path: PathBufresolver: &'a mut dyn Resolvercurrent_expansion: ExpansionDataexpansions: FxHashMap<Span, Vec<String>>
Methods

impl<'a> ExtCtxt<'a>

pub fn new(
parse_sess: &'a ParseSess,
ecfg: ExpansionConfig<'a>,
resolver: &'a mut dyn Resolver
) -> ExtCtxt<'a>

pub fn expander<'b>(&'b mut self) -> MacroExpander<'b, 'a>

Returns a Folder for deeply expanding all macros in an AST node.

pub fn monotonic_expander<'b>(&'b mut self) -> MacroExpander<'b, 'a>

Returns a Folder that deeply expands all macros and assigns all NodeIds in an AST node. Once NodeIds are assigned, the node may not be expanded, removed, or otherwise modified.

pub fn new_parser_from_tts(&self, tts: &[TokenTree]) -> Parser<'a>

pub fn source_map(&self) -> &'a SourceMap

pub fn parse_sess(&self) -> &'a ParseSess

pub fn cfg(&self) -> &CrateConfig

pub fn call_site(&self) -> Span

pub fn backtrace(&self) -> SyntaxContext

pub fn expansion_cause(&self) -> Option<Span>

Returns span for the macro which originally caused the current expansion to happen.

Stops backtracing at include! boundary.

pub fn struct_span_warn<S: Into<MultiSpan>>(
&self,
sp: S,
msg: &str
) -> DiagnosticBuilder<'a>

pub fn struct_span_err<S: Into<MultiSpan>>(
&self,
sp: S,
msg: &str
) -> DiagnosticBuilder<'a>

pub fn struct_span_fatal<S: Into<MultiSpan>>(
&self,
sp: S,
msg: &str
) -> DiagnosticBuilder<'a>

pub fn span_fatal<S: Into<MultiSpan>>(&self, sp: S, msg: &str) -> !

Emit msg attached to sp, and stop compilation immediately.

span_err should be strongly preferred where-ever possible: this should only be used when:

continuing has a high risk of flow-on errors (e.g., errors in declaring a macro would cause all uses of that macro to complain about "undefined macro"), or

there is literally nothing else that can be done (however, in most cases one can construct a dummy expression/item to substitute; we never hit resolve/type-checking so the dummy value doesn't have to match anything)

pub fn span_err<S: Into<MultiSpan>>(&self, sp: S, msg: &str)

Emit msg attached to sp, without immediately stopping compilation.

Compilation will be stopped in the near future (at the end of the macro expansion phase).

pub fn span_err_with_code<S: Into<MultiSpan>>(
&self,
sp: S,
msg: &str,
code: DiagnosticId
)

pub fn mut_span_err<S: Into<MultiSpan>>(
&self,
sp: S,
msg: &str
) -> DiagnosticBuilder<'a>

pub fn span_warn<S: Into<MultiSpan>>(&self, sp: S, msg: &str)

pub fn span_unimpl<S: Into<MultiSpan>>(&self, sp: S, msg: &str) -> !

pub fn span_bug<S: Into<MultiSpan>>(&self, sp: S, msg: &str) -> !

pub fn trace_macros_diag(&mut self)

pub fn bug(&self, msg: &str) -> !

pub fn trace_macros(&self) -> bool

pub fn set_trace_macros(&mut self, x: bool)

pub fn ident_of(&self, st: &str) -> Ident

pub fn std_path(&self, components: &[&str]) -> Vec<Ident>

pub fn name_of(&self, st: &str) -> Name

pub fn check_unused_macros(&self)