Trait syntax::ext::base::Resolver

pub trait Resolver {
    fn next_node_id(&mut self) -> NodeId;

    fn get_module_scope(&mut self, id: NodeId) -> Mark;

    fn eliminate_crate_var(&mut self, item: P<Item>) -> P<Item>;

    fn is_whitelisted_legacy_custom_derive(&self, name: Name) -> bool;

    fn visit_expansion(
        &mut self, 
        mark: Mark, 
        expansion: &Expansion, 
        derives: &[Mark]
    );

    fn add_builtin(&mut self, ident: Ident, ext: Rc<SyntaxExtension>);

    fn resolve_imports(&mut self);

    fn find_legacy_attr_invoc(
        &mut self, 
        attrs: &mut Vec<Attribute>
    ) -> Option<Attribute>;

    fn resolve_invoc(
        &mut self, 
        invoc: &mut Invocation, 
        scope: Mark, 
        force: bool
    ) -> Result<Option<Rc<SyntaxExtension>>, Determinacy>;

    fn resolve_macro(
        &mut self, 
        scope: Mark, 
        path: &Path, 
        kind: MacroKind, 
        force: bool
    ) -> Result<Rc<SyntaxExtension>, Determinacy>;

    fn check_unused_macros(&self);
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn next_node_id(&mut self) -> NodeId

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_module_scope(&mut self, id: NodeId) -> Mark

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn eliminate_crate_var(&mut self, item: P<Item>) -> P<Item>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_whitelisted_legacy_custom_derive(&self, name: Name) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_expansion(
&mut self,
mark: Mark,
expansion: &Expansion,
derives: &[Mark]
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn add_builtin(&mut self, ident: Ident, ext: Rc<SyntaxExtension>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn resolve_imports(&mut self)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn find_legacy_attr_invoc(
&mut self,
attrs: &mut Vec<Attribute>
) -> Option<Attribute>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn resolve_invoc(
&mut self,
invoc: &mut Invocation,
scope: Mark,
force: bool
) -> Result<Option<Rc<SyntaxExtension>>, Determinacy>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn resolve_macro(
&mut self,
scope: Mark,
path: &Path,
kind: MacroKind,
force: bool
) -> Result<Rc<SyntaxExtension>, Determinacy>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_unused_macros(&self)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?