Function syntax::ext::derive::add_derived_markers

pub fn add_derived_markers<T>(
    cx: &mut ExtCtxt, 
    span: Span, 
    traits: &[Path], 
    item: T
) -> T where
    T: HasAttrs, 

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?