Struct syntax::ext::deriving::generic::FieldInfo

pub struct FieldInfo<'a> {
    pub span: Span,
    pub name: Option<Ident>,
    pub self_: P<Expr>,
    pub other: Vec<P<Expr>>,
    pub attrs: &'a [Attribute],
}

Unstable

Summary of the relevant parts of a struct/enum field.

Fields

span
Unstable

name
Unstable

None for tuple structs/normal enum variants, Some for normal structs/struct enum variants.

self_
Unstable

The expression corresponding to this field of self (specifically, a reference to it).

other
Unstable

The expressions corresponding to references to this field in the other Self arguments.

attrs
Unstable

The attributes on the field