Module syntax::ext::expand

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

ExpansionConfig 	 [
Experimental
]

Invocation 	 [
Experimental
]

MacroExpander 	 [
Experimental
]

Marker 	 [
Experimental
]

Enums

Expansion 	 [
Experimental
]

ExpansionKind 	 [
Experimental
]

InvocationKind 	 [
Experimental
]

Functions

find_attr_invoc 	 [
Experimental
]