Enum syntax::ext::expand::AstFragment

pub enum AstFragment {
    OptExpr(Option<P<Expr>>),
    Expr(P<Expr>),
    Pat(P<Pat>),
    Ty(P<Ty>),
    Stmts(SmallVec<[Stmt; 1]>),
    Items(SmallVec<[P<Item>; 1]>),
    TraitItems(SmallVec<[TraitItem; 1]>),
    ImplItems(SmallVec<[ImplItem; 1]>),
    ForeignItems(SmallVec<[ForeignItem; 1]>),
}

A fragment of AST that can be produced by a single macro expansion. Can also serve as an input and intermediate result for macro expansion operations.

Variants
OptExpr(Option<P<Expr>>)Expr(P<Expr>)Pat(P<Pat>)Ty(P<Ty>)Stmts(SmallVec<[Stmt; 1]>)Items(SmallVec<[P<Item>; 1]>)TraitItems(SmallVec<[TraitItem; 1]>)ImplItems(SmallVec<[ImplItem; 1]>)ForeignItems(SmallVec<[ForeignItem; 1]>)
Methods

impl AstFragment

pub fn make_opt_expr(self) -> Option<P<Expr>>

pub fn make_expr(self) -> P<Expr>

pub fn make_pat(self) -> P<Pat>

pub fn make_ty(self) -> P<Ty>

pub fn make_stmts(self) -> SmallVec<[Stmt; 1]>

pub fn make_items(self) -> SmallVec<[P<Item>; 1]>

pub fn make_trait_items(self) -> SmallVec<[TraitItem; 1]>

pub fn make_impl_items(self) -> SmallVec<[ImplItem; 1]>

pub fn make_foreign_items(self) -> SmallVec<[ForeignItem; 1]>

pub fn mut_visit_with<F: MutVisitor>(&mut self, vis: &mut F)

pub fn visit_with<'a, V: Visitor<'a>>(&'a self, visitor: &mut V)