Module syntax::ext::hygiene

Unstable (rustc_private)

Machinery for hygienic macros, inspired by the MTWT[1] paper.

[1] Matthew Flatt, Ryan Culpepper, David Darais, and Robert Bruce Findler. 2012. Macros that work together: Compile-time bindings, partial expansion, and definition contexts. J. Funct. Program. 22, 2 (March 2012), 181-216. DOI=10.1017/S0956796812000093 http://dx.doi.org/10.1017/S0956796812000093

Structs

Mark 	 [Unstable]
A mark is a unique id associated with a macro expansion.

SyntaxContext 	 [Unstable]
A SyntaxContext represents a chain of macro expansions (represented by marks).

SyntaxContextData 	 [Unstable]

Functions

reset_hygiene_data 	 [Unstable]