Enum syntax::ext::mtwt::SyntaxContext_

pub enum SyntaxContext_ {
    EmptyCtxt,
    Mark(Mrk, SyntaxContext),
    Rename(Ident, Name, SyntaxContext),
    IllegalCtxt,
}

Unstable (rustc_private)

Variants

EmptyCtxt
Unstable (rustc_private)

Mark
Unstable (rustc_private)

Rename
Unstable (rustc_private)

flattening the name and syntaxcontext into the rename... HIDDEN INVARIANTS: 1) the first name in a Rename node can only be a programmer-supplied name. 2) Every Rename node with a given Name in the "to" slot must have the same name and context in the "from" slot. In essence, they're all pointers to a single "rename" event node.

IllegalCtxt
Unstable (rustc_private)

actually, IllegalCtxt may not be necessary.