Struct syntax::ext::placeholders::PlaceholderExpander

pub struct PlaceholderExpander<'a, 'b: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'b> PlaceholderExpander<'a, 'b>

pub fn new(cx: &'a mut ExtCtxt<'b>, monotonic: bool) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn add(&mut self, id: NodeId, expansion: Expansion, derives: Vec<Mark>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?