Module syntax::ext::proc_macro_shim::prelude

This is an experimental API. (rustc_private)

Reexports

pub use super::build_block_emitter;

pub use ast::Ident;

pub use ext::base::ExtCtxt;

pub use ext::base::MacResult;

pub use parse::token;

pub use parse::token::Token;

pub use parse::token::DelimToken;

pub use symbol::keywords;

pub use tokenstream::TokenTree;

pub use tokenstream::TokenStream;

Structs

Span 	 [

This is an experimental API.

]
Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source. You must be careful if the span crosses more than one file - you will not be able to use many of the functions on spans in codemap and you cannot assume that the length of the span = hi - lo; there may be space in the BytePos range between files.

Constants

DUMMY_SP 	 [

This is an experimental API.

]