Struct syntax::ext::quote::rt::BytePos

pub struct BytePos(pub u32);

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A byte offset. Keep this small (currently 32-bits), as AST contains a lot of them.