Trait syntax::ext::quote::rt::ExtParseUtils

pub trait ExtParseUtils {
    fn parse_item(&self, s: String) -> P<Item>;

    fn parse_expr(&self, s: String) -> P<Expr>;

    fn parse_stmt(&self, s: String) -> Stmt;

    fn parse_tts(&self, s: String) -> Vec<TokenTree>;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn parse_item(&self, s: String) -> P<Item>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn parse_expr(&self, s: String) -> P<Expr>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn parse_stmt(&self, s: String) -> Stmt

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn parse_tts(&self, s: String) -> Vec<TokenTree>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?