Enum syntax::ext::tt::macro_parser::ParseResult

pub enum ParseResult<T> {
    Success(T),
    Failure(Span, Token),
    Error(Span, String),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
Success(T)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Failure(Span, Token)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Arm failed to match. If the second parameter is token::Eof, it indicates an unexpected end of macro invocation. Otherwise, it indicates that no rules expected the given token.

Error(Span, String)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fatal error (malformed macro?). Abort compilation.