Struct syntax::ext::tt::macro_rules::ParserAnyMacro

pub struct ParserAnyMacro<'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a> ParserAnyMacro<'a>

pub fn make(self: Box<ParserAnyMacro<'a>>, kind: ExpansionKind) -> Expansion

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?