Module syntax::ext::tt::quoted

Structs

Delimited
Contains the sub-token-trees of a "delimited" token tree, such as the contents of (. Note that the delimiter itself might be NoDelim.

SequenceRepetition

Enums

KleeneOp
A Kleene-style repetition operator for token sequences.

TokenTree
Similar to tokenstream::TokenTree, except that $i, $i:ident, and $(...) are "first-class" token trees. Useful for parsing macros.

Functions

parse
Takes a tokenstream::TokenStream and returns a Vec<self::TokenTree>. Specifically, this takes a generic TokenStream, such as is used in the rest of the compiler, and returns a collection of TokenTree for use in parsing a macro.