Function syntax::ext::tt::transcribe::transcribe

pub fn transcribe(
    cx: &ExtCtxt, 
    interp: Option<HashMap<Ident, Rc<NamedMatch>>>, 
    src: Vec<TokenTree>
) -> TokenStream

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This can do Macro-By-Example transcription. On the other hand, if src contains no TokenTree::{Sequence, MetaVar, MetaVarDecl}s, interp can (and should) be None.