Enum syntax::feature_gate::AttributeGate

pub enum AttributeGate {
    Gated(&'static str, &'static str, fn(&Features) -> bool),
    Ungated,
}

Unstable (rustc_private)

Variants
Gated(&'static str, &'static str, fn(&Features) -> bool)
Unstable (rustc_private)

Is gated by a given feature gate, reason and function to check if enabled

Ungated
Unstable (rustc_private)

Ungated attribute, can be used on all release channels