Module syntax::fold

Unstable (rustc_private)

A Folder represents an AST->AST fold; it accepts an AST piece, and returns a piece of the same type. So, for instance, macro expansion is a Folder that walks over an AST and produces another AST.

Note: using a Folder (other than the MacroExpander Folder) on an AST before macro expansion is probably a bad idea. For instance, a folder renaming item names in a module will miss all of those that are created by the expansion of a macro.

Traits

Folder 	 [Unstable]

Functions

fold_attrs 	 [Unstable]

fold_thin_attrs 	 [Unstable]

noop_fold_angle_bracketed_parameter_data 	 [Unstable]

noop_fold_arg 	 [Unstable]

noop_fold_arm 	 [Unstable]

noop_fold_attribute 	 [Unstable]

noop_fold_block 	 [Unstable]

noop_fold_crate 	 [Unstable]

noop_fold_decl 	 [Unstable]

noop_fold_explicit_self 	 [Unstable]

noop_fold_explicit_self_kind 	 [Unstable]

noop_fold_expr 	 [Unstable]

noop_fold_exprs 	 [Unstable]

noop_fold_field 	 [Unstable]

noop_fold_fn_decl 	 [Unstable]

noop_fold_foreign_item 	 [Unstable]

noop_fold_foreign_mod 	 [Unstable]

noop_fold_generics 	 [Unstable]

noop_fold_ident 	 [Unstable]

noop_fold_impl_item 	 [Unstable]

noop_fold_interpolated 	 [Unstable]
apply folder to elements of interpolated nodes

noop_fold_item 	 [Unstable]

noop_fold_item_kind 	 [Unstable]

noop_fold_item_simple 	 [Unstable]

noop_fold_lifetime 	 [Unstable]

noop_fold_lifetime_def 	 [Unstable]

noop_fold_lifetime_defs 	 [Unstable]

noop_fold_lifetimes 	 [Unstable]

noop_fold_local 	 [Unstable]

noop_fold_mac 	 [Unstable]

noop_fold_meta_item 	 [Unstable]

noop_fold_meta_items 	 [Unstable]

noop_fold_method_sig 	 [Unstable]

noop_fold_mod 	 [Unstable]

noop_fold_mt 	 [Unstable]

noop_fold_opt_bounds 	 [Unstable]

noop_fold_opt_expr 	 [Unstable]

noop_fold_opt_lifetime 	 [Unstable]

noop_fold_parenthesized_parameter_data 	 [Unstable]

noop_fold_pat 	 [Unstable]

noop_fold_path 	 [Unstable]

noop_fold_path_parameters 	 [Unstable]

noop_fold_poly_trait_ref 	 [Unstable]

noop_fold_stmt 	 [Unstable]

noop_fold_struct_field 	 [Unstable]

noop_fold_token 	 [Unstable]

noop_fold_trait_item 	 [Unstable]

noop_fold_trait_ref 	 [Unstable]

noop_fold_tt 	 [Unstable]

noop_fold_tts 	 [Unstable]

noop_fold_ty 	 [Unstable]

noop_fold_ty_binding 	 [Unstable]

noop_fold_ty_param 	 [Unstable]

noop_fold_ty_param_bound 	 [Unstable]

noop_fold_ty_params 	 [Unstable]

noop_fold_usize 	 [Unstable]

noop_fold_variant 	 [Unstable]

noop_fold_variant_data 	 [Unstable]

noop_fold_view_path 	 [Unstable]

noop_fold_vis 	 [Unstable]

noop_fold_where_clause 	 [Unstable]

noop_fold_where_predicate 	 [Unstable]