Trait syntax::fold::Folder

pub trait Folder: Sized {
    fn fold_crate(&mut self, c: Crate) -> Crate { ... }

    fn fold_meta_items(&mut self, meta_items: Vec<MetaItem>) -> Vec<MetaItem> { ... }

    fn fold_meta_list_item(
        &mut self, 
        list_item: NestedMetaItem
    ) -> NestedMetaItem { ... }

    fn fold_meta_item(&mut self, meta_item: MetaItem) -> MetaItem { ... }

    fn fold_view_path(&mut self, view_path: P<ViewPath>) -> P<ViewPath> { ... }

    fn fold_foreign_item(&mut self, ni: ForeignItem) -> ForeignItem { ... }

    fn fold_item(&mut self, i: P<Item>) -> SmallVector<P<Item>> { ... }

    fn fold_item_simple(&mut self, i: Item) -> Item { ... }

    fn fold_struct_field(&mut self, sf: StructField) -> StructField { ... }

    fn fold_item_kind(&mut self, i: ItemKind) -> ItemKind { ... }

    fn fold_trait_item(&mut self, i: TraitItem) -> SmallVector<TraitItem> { ... }

    fn fold_impl_item(&mut self, i: ImplItem) -> SmallVector<ImplItem> { ... }

    fn fold_fn_decl(&mut self, d: P<FnDecl>) -> P<FnDecl> { ... }

    fn fold_block(&mut self, b: P<Block>) -> P<Block> { ... }

    fn fold_stmt(&mut self, s: Stmt) -> SmallVector<Stmt> { ... }

    fn fold_arm(&mut self, a: Arm) -> Arm { ... }

    fn fold_pat(&mut self, p: P<Pat>) -> P<Pat> { ... }

    fn fold_expr(&mut self, e: P<Expr>) -> P<Expr> { ... }

    fn fold_range_end(&mut self, re: RangeEnd) -> RangeEnd { ... }

    fn fold_opt_expr(&mut self, e: P<Expr>) -> Option<P<Expr>> { ... }

    fn fold_exprs(&mut self, es: Vec<P<Expr>>) -> Vec<P<Expr>> { ... }

    fn fold_ty(&mut self, t: P<Ty>) -> P<Ty> { ... }

    fn fold_ty_binding(&mut self, t: TypeBinding) -> TypeBinding { ... }

    fn fold_mod(&mut self, m: Mod) -> Mod { ... }

    fn fold_foreign_mod(&mut self, nm: ForeignMod) -> ForeignMod { ... }

    fn fold_global_asm(&mut self, ga: P<GlobalAsm>) -> P<GlobalAsm> { ... }

    fn fold_variant(&mut self, v: Variant) -> Variant { ... }

    fn fold_ident(&mut self, i: Ident) -> Ident { ... }

    fn fold_usize(&mut self, i: usize) -> usize { ... }

    fn fold_path(&mut self, p: Path) -> Path { ... }

    fn fold_path_parameters(&mut self, p: PathParameters) -> PathParameters { ... }

    fn fold_angle_bracketed_parameter_data(
        &mut self, 
        p: AngleBracketedParameterData
    ) -> AngleBracketedParameterData { ... }

    fn fold_parenthesized_parameter_data(
        &mut self, 
        p: ParenthesizedParameterData
    ) -> ParenthesizedParameterData { ... }

    fn fold_local(&mut self, l: P<Local>) -> P<Local> { ... }

    fn fold_mac(&mut self, _mac: Mac) -> Mac { ... }

    fn fold_macro_def(&mut self, def: MacroDef) -> MacroDef { ... }

    fn fold_lifetime(&mut self, l: Lifetime) -> Lifetime { ... }

    fn fold_lifetime_def(&mut self, l: LifetimeDef) -> LifetimeDef { ... }

    fn fold_attribute(&mut self, at: Attribute) -> Option<Attribute> { ... }

    fn fold_arg(&mut self, a: Arg) -> Arg { ... }

    fn fold_generics(&mut self, generics: Generics) -> Generics { ... }

    fn fold_trait_ref(&mut self, p: TraitRef) -> TraitRef { ... }

    fn fold_poly_trait_ref(&mut self, p: PolyTraitRef) -> PolyTraitRef { ... }

    fn fold_variant_data(&mut self, vdata: VariantData) -> VariantData { ... }

    fn fold_lifetimes(&mut self, lts: Vec<Lifetime>) -> Vec<Lifetime> { ... }

    fn fold_lifetime_defs(&mut self, lts: Vec<LifetimeDef>) -> Vec<LifetimeDef> { ... }

    fn fold_ty_param(&mut self, tp: TyParam) -> TyParam { ... }

    fn fold_ty_params(&mut self, tps: Vec<TyParam>) -> Vec<TyParam> { ... }

    fn fold_tt(&mut self, tt: TokenTree) -> TokenTree { ... }

    fn fold_tts(&mut self, tts: TokenStream) -> TokenStream { ... }

    fn fold_token(&mut self, t: Token) -> Token { ... }

    fn fold_interpolated(&mut self, nt: Nonterminal) -> Nonterminal { ... }

    fn fold_opt_lifetime(&mut self, o_lt: Option<Lifetime>) -> Option<Lifetime> { ... }

    fn fold_opt_bounds(
        &mut self, 
        b: Option<TyParamBounds>
    ) -> Option<TyParamBounds> { ... }

    fn fold_bounds(&mut self, b: TyParamBounds) -> TyParamBounds { ... }

    fn fold_ty_param_bound(&mut self, tpb: TyParamBound) -> TyParamBound { ... }

    fn fold_mt(&mut self, mt: MutTy) -> MutTy { ... }

    fn fold_field(&mut self, field: Field) -> Field { ... }

    fn fold_where_clause(&mut self, where_clause: WhereClause) -> WhereClause { ... }

    fn fold_where_predicate(
        &mut self, 
        where_predicate: WherePredicate
    ) -> WherePredicate { ... }

    fn fold_vis(&mut self, vis: Visibility) -> Visibility { ... }

    fn new_id(&mut self, i: NodeId) -> NodeId { ... }

    fn new_span(&mut self, sp: Span) -> Span { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn fold_crate(&mut self, c: Crate) -> Crate

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_meta_items(&mut self, meta_items: Vec<MetaItem>) -> Vec<MetaItem>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_meta_list_item(&mut self, list_item: NestedMetaItem) -> NestedMetaItem

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_meta_item(&mut self, meta_item: MetaItem) -> MetaItem

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_view_path(&mut self, view_path: P<ViewPath>) -> P<ViewPath>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_foreign_item(&mut self, ni: ForeignItem) -> ForeignItem

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_item(&mut self, i: P<Item>) -> SmallVector<P<Item>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_item_simple(&mut self, i: Item) -> Item

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_struct_field(&mut self, sf: StructField) -> StructField

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_item_kind(&mut self, i: ItemKind) -> ItemKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_trait_item(&mut self, i: TraitItem) -> SmallVector<TraitItem>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_impl_item(&mut self, i: ImplItem) -> SmallVector<ImplItem>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_fn_decl(&mut self, d: P<FnDecl>) -> P<FnDecl>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_block(&mut self, b: P<Block>) -> P<Block>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_stmt(&mut self, s: Stmt) -> SmallVector<Stmt>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_arm(&mut self, a: Arm) -> Arm

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_pat(&mut self, p: P<Pat>) -> P<Pat>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_expr(&mut self, e: P<Expr>) -> P<Expr>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_range_end(&mut self, re: RangeEnd) -> RangeEnd

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_opt_expr(&mut self, e: P<Expr>) -> Option<P<Expr>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_exprs(&mut self, es: Vec<P<Expr>>) -> Vec<P<Expr>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_ty(&mut self, t: P<Ty>) -> P<Ty>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_ty_binding(&mut self, t: TypeBinding) -> TypeBinding

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_mod(&mut self, m: Mod) -> Mod

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_foreign_mod(&mut self, nm: ForeignMod) -> ForeignMod

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_global_asm(&mut self, ga: P<GlobalAsm>) -> P<GlobalAsm>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_variant(&mut self, v: Variant) -> Variant

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_ident(&mut self, i: Ident) -> Ident

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_usize(&mut self, i: usize) -> usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_path(&mut self, p: Path) -> Path

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_path_parameters(&mut self, p: PathParameters) -> PathParameters

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_angle_bracketed_parameter_data(
&mut self,
p: AngleBracketedParameterData
) -> AngleBracketedParameterData

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_parenthesized_parameter_data(
&mut self,
p: ParenthesizedParameterData
) -> ParenthesizedParameterData

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_local(&mut self, l: P<Local>) -> P<Local>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_mac(&mut self, _mac: Mac) -> Mac

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_macro_def(&mut self, def: MacroDef) -> MacroDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_lifetime(&mut self, l: Lifetime) -> Lifetime

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_lifetime_def(&mut self, l: LifetimeDef) -> LifetimeDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_attribute(&mut self, at: Attribute) -> Option<Attribute>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_arg(&mut self, a: Arg) -> Arg

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_generics(&mut self, generics: Generics) -> Generics

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_trait_ref(&mut self, p: TraitRef) -> TraitRef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_poly_trait_ref(&mut self, p: PolyTraitRef) -> PolyTraitRef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_variant_data(&mut self, vdata: VariantData) -> VariantData

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_lifetimes(&mut self, lts: Vec<Lifetime>) -> Vec<Lifetime>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_lifetime_defs(&mut self, lts: Vec<LifetimeDef>) -> Vec<LifetimeDef>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_ty_param(&mut self, tp: TyParam) -> TyParam

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_ty_params(&mut self, tps: Vec<TyParam>) -> Vec<TyParam>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_tt(&mut self, tt: TokenTree) -> TokenTree

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_tts(&mut self, tts: TokenStream) -> TokenStream

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_token(&mut self, t: Token) -> Token

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_interpolated(&mut self, nt: Nonterminal) -> Nonterminal

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_opt_lifetime(&mut self, o_lt: Option<Lifetime>) -> Option<Lifetime>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_opt_bounds(&mut self, b: Option<TyParamBounds>) -> Option<TyParamBounds>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_bounds(&mut self, b: TyParamBounds) -> TyParamBounds

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_ty_param_bound(&mut self, tpb: TyParamBound) -> TyParamBound

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_mt(&mut self, mt: MutTy) -> MutTy

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_field(&mut self, field: Field) -> Field

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_where_clause(&mut self, where_clause: WhereClause) -> WhereClause

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_where_predicate(
&mut self,
where_predicate: WherePredicate
) -> WherePredicate

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_vis(&mut self, vis: Visibility) -> Visibility

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn new_id(&mut self, i: NodeId) -> NodeId

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn new_span(&mut self, sp: Span) -> Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?