Module syntax::generic

Modules

def 🔒

fold 🔒

map 🔒

visit 🔒

Structs

Case

Cocase

Codata

Codef

Comatch

Const

Ctor

Data

Decls

Def

Dtor

Id

Id

Impl

Match

Param

Prg

Telescope

Wrapper type signifying the wrapped parameters have telescope semantics. I.e. each parameter binding in the parameter list is in scope for the following parameters.

TypAbs

TypApp

Enums

Decl

Exp

Traits

Fold

Folder

HasPhase

Map

Mapper

Out

Phase

Visit

Visitor

Type Definitions

Args

Params