Struct syntax::leveled_ctx::LeveledCtx

pub struct LeveledCtx {
    bound: Vec<usize>,
}

Fields
bound: Vec<usize>

Number of binders in the second dimension for each first dimension

Implementations

impl LeveledCtx

pub fn empty() -> Self

pub fn bind<T, I, O, F>(&mut self, iter: I, f: F) -> O where
I: Iterator<Item = T>,
F: FnOnce(&mut LeveledCtx) -> O,

Bind an iterator iter of binders

pub fn bind_fold<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> O2 where
F1: Fn(&mut LeveledCtx, O1, T) -> O1,
F2: FnOnce(&mut LeveledCtx, O1) -> O2,

Bind an iterator iter of binders

Fold the iterator and consume the result under the inner context with all binders in scope.

iter - An iterator of binders implementing Named.

acc - Accumulator for folding the iterator

f_acc - Accumulator function run for each binder

f_inner - Inner function computing the final result under the context of all binders

fn level_inc_fst(&mut self)

Increment the first component of the current De-Bruijn level

fn level_dec_fst(&mut self)

Decrement the first component of the current De-Bruijn level

fn push(&mut self)

Push a binder contained in a binder list, incrementing the second dimension of the current De Bruijn level

fn pop(&mut self)

Push a binder contained in a binder list, decrementing the second dimension of the current De Bruijn level