Trait syntax::parse::obsolete::ParserObsoleteMethods

pub trait ParserObsoleteMethods {
    fn obsolete(&mut self, sp: Span, kind: ObsoleteSyntax);

    fn report(
        &mut self, 
        sp: Span, 
        kind: ObsoleteSyntax, 
        kind_str: &str, 
        desc: &str, 
        error: bool
    );
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn obsolete(&mut self, sp: Span, kind: ObsoleteSyntax)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reports an obsolete syntax non-fatal error.

fn report(
&mut self,
sp: Span,
kind: ObsoleteSyntax,
kind_str: &str,
desc: &str,
error: bool
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?