Module syntax::parse::parser

Unstable

Reexports

pub use self::PathParsingMode::*;

Structs

Parser 	 [Unstable]

Restrictions 	 [Unstable]

Enums

BoundParsingMode 	 [Unstable]
How to parse a bound, whether to allow bound modifiers such as ?.

PathParsingMode 	 [Unstable]
How to parse a path. There are four different kinds of paths, all of which are parsed somewhat differently.

TokenType 	 [Unstable]