Struct syntax::parse::parser::Parser

pub struct Parser<'a> {
    pub sess: &'a ParseSess,
    pub token: Token,
    pub span: Span,
    pub last_span: Span,
    pub cfg: CrateConfig,
    pub buffer: [TokenAndSpan; 4],
    pub buffer_start: isize,
    pub buffer_end: isize,
    pub tokens_consumed: usize,
    pub restrictions: Restrictions,
    pub quote_depth: usize,
    pub reader: Box<Reader + 'a>,
    pub obsolete_set: HashSet<ObsoleteSyntax>,
    pub directory: PathBuf,
    pub open_braces: Vec<(DelimToken, Span)>,
    pub owns_directory: bool,
    pub root_module_name: Option<String>,
    pub expected_tokens: Vec<TokenType>,
    // some fields omitted
}

Unstable (rustc_private)

Fields
sess: &'a ParseSess
Unstable (rustc_private)
token: Token
Unstable (rustc_private)

the current token:

span: Span
Unstable (rustc_private)

the span of the current token:

last_span: Span
Unstable (rustc_private)

the span of the prior token:

cfg: CrateConfig
Unstable (rustc_private)
buffer: [TokenAndSpan; 4]
Unstable (rustc_private)
buffer_start: isize
Unstable (rustc_private)
buffer_end: isize
Unstable (rustc_private)
tokens_consumed: usize
Unstable (rustc_private)
restrictions: Restrictions
Unstable (rustc_private)
quote_depth: usize
Unstable (rustc_private)
reader: Box<Reader + 'a>
Unstable (rustc_private)
obsolete_set: HashSet<ObsoleteSyntax>
Unstable (rustc_private)

The set of seen errors about obsolete syntax. Used to suppress extra detail when the same error is seen twice

directory: PathBuf
Unstable (rustc_private)

Used to determine the path to externally loaded source files

open_braces: Vec<(DelimToken, Span)>
Unstable (rustc_private)

Stack of open delimiters and their spans. Used for error message.

owns_directory: bool
Unstable (rustc_private)

Flag if this parser "owns" the directory that it is currently parsing in. This will affect how nested files are looked up.

root_module_name: Option<String>
Unstable (rustc_private)

Name of the root module this parser originated from. If None, then the name is not known. This does not change while the parser is descending into modules, and sub-parsers have new values for this name.

expected_tokens: Vec<TokenType>
Unstable (rustc_private)

Methods

impl<'a> Parser<'a>

fn new(sess: &'a ParseSess, cfg: CrateConfig, rdr: Box<Reader + 'a>) -> Parser<'a>

Unstable (rustc_private)

fn token_to_string(token: &Token) -> String

Unstable (rustc_private)

Convert a token to a string using self's reader

fn this_token_to_string(&self) -> String

Unstable (rustc_private)

Convert the current token to a string using self's reader

fn this_token_descr(&self) -> String

Unstable (rustc_private)

fn unexpected_last<T>(&self, t: &Token) -> PResult<'a, T>

Unstable (rustc_private)

fn unexpected<T>(&mut self) -> PResult<'a, T>

Unstable (rustc_private)

fn expect(&mut self, t: &Token) -> PResult<'a, ()>

Unstable (rustc_private)

Expect and consume the token t. Signal an error if the next token is not t.

fn expect_one_of(&mut self, edible: &[Token], inedible: &[Token]) -> PResult<'a, ()>

Unstable (rustc_private)

Expect next token to be edible or inedible token. If edible, then consume it; if inedible, then return without consuming anything. Signal a fatal error if next token is unexpected.

fn parse_ident(&mut self) -> PResult<'a, Ident>

Unstable (rustc_private)

fn check(&mut self, tok: &Token) -> bool

Unstable (rustc_private)

Check if the next token is tok, and return true if so.

This method will automatically add tok to expected_tokens if tok is not encountered.

fn eat(&mut self, tok: &Token) -> bool

Unstable (rustc_private)

Consume token 'tok' if it exists. Returns true if the given token was present, false otherwise.

fn check_keyword(&mut self, kw: Keyword) -> bool

Unstable (rustc_private)

fn eat_keyword(&mut self, kw: Keyword) -> bool

Unstable (rustc_private)

If the next token is the given keyword, eat it and return true. Otherwise, return false.

fn eat_keyword_noexpect(&mut self, kw: Keyword) -> bool

Unstable (rustc_private)

fn check_contextual_keyword(&mut self, ident: Ident) -> bool

Unstable (rustc_private)

fn eat_contextual_keyword(&mut self, ident: Ident) -> bool

Unstable (rustc_private)

fn expect_keyword(&mut self, kw: Keyword) -> PResult<'a, ()>

Unstable (rustc_private)

If the given word is not a keyword, signal an error. If the next token is not the given word, signal an error. Otherwise, eat it.

fn check_strict_keywords(&mut self)

Unstable (rustc_private)

Signal an error if the given string is a strict keyword

fn check_reserved_keywords(&mut self)

Unstable (rustc_private)

Signal an error if the current token is a reserved keyword

fn expect_no_suffix(&self, sp: Span, kind: &str, suffix: Option<Name>)

Unstable (rustc_private)

fn expect_gt(&mut self) -> PResult<'a, ()>

Unstable (rustc_private)

Expect and consume a GT. if a >> is seen, replace it with a single > and continue. If a GT is not seen, signal an error.

fn parse_seq_to_before_gt_or_return<T, F>(&mut self, sep: Option<Token>, f: F) -> PResult<'a, (P<[T]>, bool)> where F: FnMut(&mut Parser<'a>) -> PResult<'a, Option<T>>

Unstable (rustc_private)

fn parse_seq_to_before_gt<T, F>(&mut self, sep: Option<Token>, f: F) -> PResult<'a, P<[T]>> where F: FnMut(&mut Parser<'a>) -> PResult<'a, T>

Unstable (rustc_private)

Parse a sequence bracketed by '<' and '>', stopping before the '>'.

fn parse_seq_to_gt<T, F>(&mut self, sep: Option<Token>, f: F) -> PResult<'a, P<[T]>> where F: FnMut(&mut Parser<'a>) -> PResult<'a, T>

Unstable (rustc_private)

fn parse_seq_to_gt_or_return<T, F>(&mut self, sep: Option<Token>, f: F) -> PResult<'a, (P<[T]>, bool)> where F: FnMut(&mut Parser<'a>) -> PResult<'a, Option<T>>

Unstable (rustc_private)

fn eat_to_tokens(&mut self, kets: &[&Token])

Unstable (rustc_private)

Eat and discard tokens until one of kets is encountered. Respects token trees, passes through any errors encountered. Used for error recovery.

fn parse_seq_to_end<T, F>(&mut self, ket: &Token, sep: SeqSep, f: F) -> PResult<'a, Vec<T>> where F: FnMut(&mut Parser<'a>) -> PResult<'a, T>

Unstable (rustc_private)

Parse a sequence, including the closing delimiter. The function f must consume tokens until reaching the next separator or closing bracket.

fn parse_seq_to_before_end<T, F>(&mut self, ket: &Token, sep: SeqSep, f: F) -> Vec<T> where F: FnMut(&mut Parser<'a>) -> PResult<'a, T>

Unstable (rustc_private)

Parse a sequence, not including the closing delimiter. The function f must consume tokens until reaching the next separator or closing bracket.

fn parse_unspanned_seq<T, F>(&mut self, bra: &Token, ket: &Token, sep: SeqSep, f: F) -> PResult<'a, Vec<T>> where F: FnMut(&mut Parser<'a>) -> PResult<'a, T>

Unstable (rustc_private)

Parse a sequence, including the closing delimiter. The function f must consume tokens until reaching the next separator or closing bracket.

fn parse_seq<T, F>(&mut self, bra: &Token, ket: &Token, sep: SeqSep, f: F) -> PResult<'a, Spanned<Vec<T>>> where F: FnMut(&mut Parser<'a>) -> PResult<'a, T>

Unstable (rustc_private)

fn bump(&mut self)

Unstable (rustc_private)

Advance the parser by one token

fn bump_and_get(&mut self) -> Token

Unstable (rustc_private)

Advance the parser by one token and return the bumped token.

fn bump_with(&mut self, next: Token, lo: BytePos, hi: BytePos)

Unstable (rustc_private)

Advance the parser using provided token as a next one. Use this when consuming a part of a token. For example a single < from <<.

fn buffer_length(&mut self) -> isize

Unstable (rustc_private)

fn look_ahead<R, F>(&mut self, distance: usize, f: F) -> R where F: FnOnce(&Token) -> R

Unstable (rustc_private)

fn fatal(&self, m: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn span_fatal(&self, sp: Span, m: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn span_fatal_help(&self, sp: Span, m: &str, help: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn bug(&self, m: &str) -> !

Unstable (rustc_private)

fn warn(&self, m: &str)

Unstable (rustc_private)

fn span_warn(&self, sp: Span, m: &str)

Unstable (rustc_private)

fn span_err(&self, sp: Span, m: &str)

Unstable (rustc_private)

fn span_err_help(&self, sp: Span, m: &str, h: &str)

Unstable (rustc_private)

fn span_bug(&self, sp: Span, m: &str) -> !

Unstable (rustc_private)

fn abort_if_errors(&self)

Unstable (rustc_private)

fn diagnostic(&self) -> &'a Handler

Unstable (rustc_private)

fn id_to_interned_str(&mut self, id: Ident) -> InternedString

Unstable (rustc_private)

fn token_is_bare_fn_keyword(&mut self) -> bool

Unstable (rustc_private)

Is the current token one of the keywords that signals a bare function type?

fn get_lifetime(&mut self) -> Ident

Unstable (rustc_private)

fn parse_for_in_type(&mut self) -> PResult<'a, TyKind>

Unstable (rustc_private)

fn parse_impl_trait_type(&mut self) -> PResult<'a, TyKind>

Unstable (rustc_private)

fn parse_ty_path(&mut self) -> PResult<'a, TyKind>

Unstable (rustc_private)

fn parse_ty_bare_fn(&mut self, lifetime_defs: Vec<LifetimeDef>) -> PResult<'a, TyKind>

Unstable (rustc_private)

parse a TyKind::BareFn type:

fn parse_unsafety(&mut self) -> PResult<'a, Unsafety>

Unstable (rustc_private)

fn parse_trait_item(&mut self) -> PResult<'a, TraitItem>

Unstable (rustc_private)

Parse the items in a trait declaration

fn parse_trait_items(&mut self) -> PResult<'a, Vec<TraitItem>>

Unstable (rustc_private)

Parse the items in a trait declaration

fn parse_mt(&mut self) -> PResult<'a, MutTy>

Unstable (rustc_private)

Parse a possibly mutable type

fn parse_ret_ty(&mut self) -> PResult<'a, FunctionRetTy>

Unstable (rustc_private)

Parse optional return type [ -> TY ] in function decl

fn parse_ty_sum(&mut self) -> PResult<'a, P<Ty>>

Unstable (rustc_private)

Parse a type in a context where T1+T2 is allowed.

fn parse_ty(&mut self) -> PResult<'a, P<Ty>>

Unstable (rustc_private)

Parse a type.

fn parse_borrowed_pointee(&mut self) -> PResult<'a, TyKind>

Unstable (rustc_private)

fn parse_ptr(&mut self) -> PResult<'a, MutTy>

Unstable (rustc_private)

fn is_named_argument(&mut self) -> bool

Unstable (rustc_private)

fn parse_arg_general(&mut self, require_name: bool) -> PResult<'a, Arg>

Unstable (rustc_private)

This version of parse arg doesn't necessarily require identifier names.

fn parse_arg(&mut self) -> PResult<'a, Arg>

Unstable (rustc_private)

Parse a single function argument

fn parse_fn_block_arg(&mut self) -> PResult<'a, Arg>

Unstable (rustc_private)

Parse an argument in a lambda header e.g. |arg, arg|

fn maybe_parse_fixed_length_of_vec(&mut self) -> PResult<'a, Option<P<Expr>>>

Unstable (rustc_private)

fn parse_lit_token(&mut self) -> PResult<'a, LitKind>

Unstable (rustc_private)

Matches token_lit = LIT_INTEGER | ...

fn parse_lit(&mut self) -> PResult<'a, Lit>

Unstable (rustc_private)

Matches lit = true | false | token_lit

fn parse_pat_literal_maybe_minus(&mut self) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

matches '-' lit | lit

fn parse_path_segment_ident(&mut self) -> PResult<'a, Ident>

Unstable (rustc_private)

fn parse_qualified_path(&mut self, mode: PathStyle) -> PResult<'a, (QSelf, Path)>

Unstable (rustc_private)

Parses qualified path.

Assumes that the leading < has been parsed already.

Qualifed paths are a part of the universal function call syntax (UFCS).

qualified_path = <type [as trait_ref]>::path

See parse_path for mode meaning.

Examples:

<T as U>::a <T as U>::F::a::<S>

fn parse_path(&mut self, mode: PathStyle) -> PResult<'a, Path>

Unstable (rustc_private)

Parses a path and optional type parameter bounds, depending on the mode. The mode parameter determines whether lifetimes, types, and/or bounds are permitted and whether :: must precede type parameter groups.

fn parse_path_segments_without_colons(&mut self) -> PResult<'a, Vec<PathSegment>>

Unstable (rustc_private)

Examples: - a::b<T,U>::c<V,W> - a::b<T,U>::c(V) -> W - a::b<T,U>::c(V)

fn parse_path_segments_with_colons(&mut self) -> PResult<'a, Vec<PathSegment>>

Unstable (rustc_private)

Examples: - a::b::<T,U>::c

fn parse_path_segments_without_types(&mut self) -> PResult<'a, Vec<PathSegment>>

Unstable (rustc_private)

Examples: - a::b::c

fn parse_opt_lifetime(&mut self) -> PResult<'a, Option<Lifetime>>

Unstable (rustc_private)

parses 0 or 1 lifetime

fn parse_lifetime(&mut self) -> PResult<'a, Lifetime>

Unstable (rustc_private)

Parses a single lifetime Matches lifetime = LIFETIME

fn parse_lifetime_defs(&mut self, followed_by_ty_params: Option<&mut Vec<Attribute>>) -> PResult<'a, Vec<LifetimeDef>>

Unstable (rustc_private)

Parses lifetime_defs = [ lifetime_defs { ',' lifetime_defs } ] where lifetime_def = lifetime [':' lifetimes]

If followed_by_ty_params is None, then we are in a context where only lifetime parameters are allowed, and thus we should error if we encounter attributes after the bound lifetimes.

If followed_by_ty_params is Some(r), then there may be type parameter bindings after the lifetimes, so we should pass along the parsed attributes to be attached to the first such type parmeter.

fn parse_lifetimes(&mut self, sep: Token) -> PResult<'a, Vec<Lifetime>>

Unstable (rustc_private)

matches lifetimes = ( lifetime ) | ( lifetime , lifetimes ) actually, it matches the empty one too, but putting that in there messes up the grammar....

Parses zero or more comma separated lifetimes. Expects each lifetime to be followed by either a comma or >. Used when parsing type parameter lists, where we expect something like <'a, 'b, T>.

fn parse_mutability(&mut self) -> PResult<'a, Mutability>

Unstable (rustc_private)

Parse mutability (mut or nothing).

fn parse_field_name(&mut self) -> PResult<'a, Ident>

Unstable (rustc_private)

fn parse_field(&mut self) -> PResult<'a, Field>

Unstable (rustc_private)

Parse ident COLON expr

fn mk_expr(&mut self, lo: BytePos, hi: BytePos, node: ExprKind, attrs: ThinVec<Attribute>) -> P<Expr>

Unstable (rustc_private)

fn mk_unary(&mut self, unop: UnOp, expr: P<Expr>) -> ExprKind

Unstable (rustc_private)

fn mk_binary(&mut self, binop: BinOp, lhs: P<Expr>, rhs: P<Expr>) -> ExprKind

Unstable (rustc_private)

fn mk_call(&mut self, f: P<Expr>, args: Vec<P<Expr>>) -> ExprKind

Unstable (rustc_private)

fn mk_index(&mut self, expr: P<Expr>, idx: P<Expr>) -> ExprKind

Unstable (rustc_private)

fn mk_range(&mut self, start: Option<P<Expr>>, end: Option<P<Expr>>, limits: RangeLimits) -> PResult<'a, ExprKind>

Unstable (rustc_private)

fn mk_field(&mut self, expr: P<Expr>, ident: SpannedIdent) -> ExprKind

Unstable (rustc_private)

fn mk_tup_field(&mut self, expr: P<Expr>, idx: Spanned<usize>) -> ExprKind

Unstable (rustc_private)

fn mk_assign_op(&mut self, binop: BinOp, lhs: P<Expr>, rhs: P<Expr>) -> ExprKind

Unstable (rustc_private)

fn mk_mac_expr(&mut self, lo: BytePos, hi: BytePos, m: Mac_, attrs: ThinVec<Attribute>) -> P<Expr>

Unstable (rustc_private)

fn mk_lit_u32(&mut self, i: u32, attrs: ThinVec<Attribute>) -> P<Expr>

Unstable (rustc_private)

fn parse_block_expr(&mut self, lo: BytePos, blk_mode: BlockCheckMode, outer_attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse a block or unsafe block

fn parse_dot_or_call_expr(&mut self, already_parsed_attrs: Option<ThinVec<Attribute>>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

parse a.b or a(13) or a[4] or just a

fn parse_dot_or_call_expr_with(&mut self, e0: P<Expr>, lo: BytePos, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

fn check_unknown_macro_variable(&mut self)

Unstable (rustc_private)

fn parse_sep_and_kleene_op(&mut self) -> PResult<'a, (Option<Token>, KleeneOp)>

Unstable (rustc_private)

Parse an optional separator followed by a Kleene-style repetition token (+ or *).

fn parse_token_tree(&mut self) -> PResult<'a, TokenTree>

Unstable (rustc_private)

parse a single token tree from the input.

fn parse_all_token_trees(&mut self) -> PResult<'a, Vec<TokenTree>>

Unstable (rustc_private)

fn parse_prefix_expr(&mut self, already_parsed_attrs: Option<ThinVec<Attribute>>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse a prefix-unary-operator expr

fn parse_assoc_expr(&mut self, already_parsed_attrs: Option<ThinVec<Attribute>>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse an associative expression

This parses an expression accounting for associativity and precedence of the operators in the expression.

fn parse_assoc_expr_with(&mut self, min_prec: usize, lhs: LhsExpr) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse an associative expression with operators of at least min_prec precedence

fn parse_if_expr(&mut self, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse an 'if' or 'if let' expression ('if' token already eaten)

fn parse_if_let_expr(&mut self, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse an 'if let' expression ('if' token already eaten)

fn parse_lambda_expr(&mut self, lo: BytePos, capture_clause: CaptureBy, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

fn parse_else_expr(&mut self) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

fn parse_for_expr(&mut self, opt_ident: Option<SpannedIdent>, span_lo: BytePos, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse a 'for' .. 'in' expression ('for' token already eaten)

fn parse_while_expr(&mut self, opt_ident: Option<SpannedIdent>, span_lo: BytePos, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse a 'while' or 'while let' expression ('while' token already eaten)

fn parse_while_let_expr(&mut self, opt_ident: Option<SpannedIdent>, span_lo: BytePos, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse a 'while let' expression ('while' token already eaten)

fn parse_loop_expr(&mut self, opt_ident: Option<SpannedIdent>, span_lo: BytePos, attrs: ThinVec<Attribute>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

fn parse_arm(&mut self) -> PResult<'a, Arm>

Unstable (rustc_private)

fn parse_expr(&mut self) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse an expression

fn with_res<F, T>(&mut self, r: Restrictions, f: F) -> T where F: FnOnce(&mut Self) -> T

Unstable (rustc_private)

Evaluate the closure with restrictions in place.

After the closure is evaluated, restrictions are reset.

fn parse_expr_res(&mut self, r: Restrictions, already_parsed_attrs: Option<ThinVec<Attribute>>) -> PResult<'a, P<Expr>>

Unstable (rustc_private)

Parse an expression, subject to the given restrictions

fn parse_pat(&mut self) -> PResult<'a, P<Pat>>

Unstable (rustc_private)

Parse a pattern.

fn parse_stmt(&mut self) -> PResult<'a, Option<Stmt>>

Unstable (rustc_private)

Parse a statement. This stops just before trailing semicolons on everything but items. e.g. a StmtKind::Semi parses to a StmtKind::Expr, leaving the trailing ; unconsumed.

fn parse_block(&mut self) -> PResult<'a, P<Block>>

Unstable (rustc_private)

Parse a block. No inner attrs are allowed.

fn parse_full_stmt(&mut self, macro_legacy_warnings: bool) -> PResult<'a, Option<Stmt>>

Unstable (rustc_private)

Parse a statement, including the trailing semicolon.

fn parse_generics(&mut self) -> PResult<'a, Generics>

Unstable (rustc_private)

Parse a set of optional generic type parameter declarations. Where clauses are not parsed here, and must be added later via parse_where_clause().

matches generics = ( ) | ( < > ) | ( < typaramseq ( , )? > ) | ( < lifetimes ( , )? > ) | ( < lifetimes , typaramseq ( , )? > ) where typaramseq = ( typaram ) | ( typaram , typaramseq )

fn parse_where_clause(&mut self) -> PResult<'a, WhereClause>

Unstable (rustc_private)

Parses an optional where clause and places it in generics.

where T : Trait<U, V> + 'b, 'a : 'bRun

fn parse_fn_decl(&mut self, allow_variadic: bool) -> PResult<'a, P<FnDecl>>

Unstable (rustc_private)

Parse the argument list and result type of a function declaration

fn is_const_item(&mut self) -> bool

Unstable (rustc_private)

true if we are looking at const ID, false for things like const fn etc

fn parse_fn_front_matter(&mut self) -> PResult<'a, (Spanned<Constness>, Unsafety, Abi)>

Unstable (rustc_private)

parses all the "front matter" for a fn declaration, up to and including the fn keyword:

const fn

unsafe fn

const unsafe fn

extern fn

etc

fn parse_impl_item(&mut self) -> PResult<'a, ImplItem>

Unstable (rustc_private)

Parse an impl item.

fn parse_record_struct_body(&mut self) -> PResult<'a, Vec<StructField>>

Unstable (rustc_private)

fn parse_tuple_struct_body(&mut self) -> PResult<'a, Vec<StructField>>

Unstable (rustc_private)

fn parse_single_struct_field(&mut self, lo: BytePos, vis: Visibility, attrs: Vec<Attribute>) -> PResult<'a, StructField>

Unstable (rustc_private)

Parse a structure field declaration

fn submod_path_from_attr(attrs: &[Attribute], dir_path: &Path) -> Option<PathBuf>

Unstable (rustc_private)

fn default_submod_path(id: Ident, dir_path: &Path, codemap: &CodeMap) -> ModulePath

Unstable (rustc_private)

Returns either a path to a module, or .

fn parse_item(&mut self) -> PResult<'a, Option<P<Item>>>

Unstable (rustc_private)

fn parse_crate_mod(&mut self) -> PResult<'a, Crate>

Unstable (rustc_private)

Parses a source module as a crate. This is the main entry point for the parser.

fn parse_optional_str(&mut self) -> Option<(InternedString, StrStyle, Option<Name>)>

Unstable (rustc_private)

fn parse_str(&mut self) -> PResult<'a, (InternedString, StrStyle)>

Unstable (rustc_private)

impl<'a> Parser<'a>

fn parse_outer_attributes(&mut self) -> PResult<'a, Vec<Attribute>>

Unstable (rustc_private)

Parse attributes that appear before an item

fn parse_attribute(&mut self, permit_inner: bool) -> PResult<'a, Attribute>

Unstable (rustc_private)

Matches attribute = # ! [ meta_item ]

If permit_inner is true, then a leading ! indicates an inner attribute

fn parse_inner_attributes(&mut self) -> PResult<'a, Vec<Attribute>>

Unstable (rustc_private)

Parse attributes that appear after the opening of an item. These should be preceded by an exclamation mark, but we accept and warn about one terminated by a semicolon. matches inner_attrs*

fn parse_meta_item(&mut self) -> PResult<'a, P<MetaItem>>

Unstable (rustc_private)

Per RFC#1559, matches the following grammar:

meta_item : IDENT ( '=' UNSUFFIXED_LIT | '(' meta_item_inner? ')' )? ; meta_item_inner : (meta_item | UNSUFFIXED_LIT) (',' meta_item_inner)? ;

impl<'a> Parser<'a>

fn parse_expansion(&mut self, kind: ExpansionKind, macro_legacy_warnings: bool) -> PResult<'a, Expansion>

Unstable (rustc_private)

fn ensure_complete_parse(&mut self, macro_name: Name, kind_name: &str, span: Span)

Unstable (rustc_private)