Module syntax::parse::token

Unstable (rustc_private)

Reexports

pub use self::BinOpToken::*;

pub use self::Nonterminal::*;

pub use self::DelimToken::*;

pub use self::Lit::*;

pub use self::Token::*;

Modules

keywords 	 [Unstable]

Structs

InternedString 	 [Unstable]
Represents a string stored in the thread-local interner. Because the interner lives for the life of the thread, this can be safely treated as an immortal string, as long as it never crosses between threads.

Enums

BinOpToken 	 [Unstable]

DelimToken 	 [Unstable]
A delimiter token

Lit 	 [Unstable]

Nonterminal 	 [Unstable]
For interpolation during macro expansion.

SpecialMacroVar 	 [Unstable]

Token 	 [Unstable]

Functions

fresh_mark 	 [Unstable]

fresh_name 	 [Unstable]

gensym 	 [Unstable]
gensym's a new usize, using the current interner.

gensym_ident 	 [Unstable]
Maps a string to a gensym'ed identifier.

get_ident_interner 	 [Unstable]

intern 	 [Unstable]
Maps a string to its interned representation.

intern_and_get_ident 	 [Unstable]
Interns and returns the string contents of an identifier, using the thread-local interner.

reset_ident_interner 	 [Unstable]
Reset the ident interner to its initial state.

str_to_ident 	 [Unstable]
Maps a string to an identifier with an empty syntax context.

Type Definitions

IdentInterner 	 [Unstable]