Enum syntax::parse::token::Token

pub enum Token {
    Eq,
    Lt,
    Le,
    EqEq,
    Ne,
    Ge,
    Gt,
    AndAnd,
    OrOr,
    Not,
    Tilde,
    BinOp(BinOpToken),
    BinOpEq(BinOpToken),
    At,
    Dot,
    DotDot,
    DotDotDot,
    Comma,
    Semi,
    Colon,
    ModSep,
    RArrow,
    LArrow,
    FatArrow,
    Pound,
    Dollar,
    Question,
    OpenDelim(DelimToken),
    CloseDelim(DelimToken),
    Literal(Lit, Option<Name>),
    Ident(Ident),
    Underscore,
    Lifetime(Ident),
    Interpolated(Rc<Nonterminal>),
    DocComment(Name),
    MatchNt(Ident, Ident),
    SubstNt(Ident),
    Whitespace,
    Comment,
    Shebang(Name),
    Eof,
}

🔬 This is a nightly-only experimental API. (rustc_private)

Variants
Eq

🔬 This is a nightly-only experimental API. (rustc_private)

Lt

🔬 This is a nightly-only experimental API. (rustc_private)

Le

🔬 This is a nightly-only experimental API. (rustc_private)

EqEq

🔬 This is a nightly-only experimental API. (rustc_private)

Ne

🔬 This is a nightly-only experimental API. (rustc_private)

Ge

🔬 This is a nightly-only experimental API. (rustc_private)

Gt

🔬 This is a nightly-only experimental API. (rustc_private)

AndAnd

🔬 This is a nightly-only experimental API. (rustc_private)

OrOr

🔬 This is a nightly-only experimental API. (rustc_private)

Not

🔬 This is a nightly-only experimental API. (rustc_private)

Tilde

🔬 This is a nightly-only experimental API. (rustc_private)

BinOp(BinOpToken)

🔬 This is a nightly-only experimental API. (rustc_private)

BinOpEq(BinOpToken)

🔬 This is a nightly-only experimental API. (rustc_private)

At

🔬 This is a nightly-only experimental API. (rustc_private)

Dot

🔬 This is a nightly-only experimental API. (rustc_private)

DotDot

🔬 This is a nightly-only experimental API. (rustc_private)

DotDotDot

🔬 This is a nightly-only experimental API. (rustc_private)

Comma

🔬 This is a nightly-only experimental API. (rustc_private)

Semi

🔬 This is a nightly-only experimental API. (rustc_private)

Colon

🔬 This is a nightly-only experimental API. (rustc_private)

ModSep

🔬 This is a nightly-only experimental API. (rustc_private)

RArrow

🔬 This is a nightly-only experimental API. (rustc_private)

LArrow

🔬 This is a nightly-only experimental API. (rustc_private)

FatArrow

🔬 This is a nightly-only experimental API. (rustc_private)

Pound

🔬 This is a nightly-only experimental API. (rustc_private)

Dollar

🔬 This is a nightly-only experimental API. (rustc_private)

Question

🔬 This is a nightly-only experimental API. (rustc_private)

OpenDelim(DelimToken)

🔬 This is a nightly-only experimental API. (rustc_private)

An opening delimiter, eg. {

CloseDelim(DelimToken)

🔬 This is a nightly-only experimental API. (rustc_private)

A closing delimiter, eg. }

Literal(Lit, Option<Name>)

🔬 This is a nightly-only experimental API. (rustc_private)

Ident(Ident)

🔬 This is a nightly-only experimental API. (rustc_private)

Underscore

🔬 This is a nightly-only experimental API. (rustc_private)

Lifetime(Ident)

🔬 This is a nightly-only experimental API. (rustc_private)

Interpolated(Rc<Nonterminal>)

🔬 This is a nightly-only experimental API. (rustc_private)

DocComment(Name)

🔬 This is a nightly-only experimental API. (rustc_private)

Doc comment

MatchNt(Ident, Ident)

🔬 This is a nightly-only experimental API. (rustc_private)

Parse a nonterminal (name to bind, name of NT)

SubstNt(Ident)

🔬 This is a nightly-only experimental API. (rustc_private)

A syntactic variable that will be filled in by macro expansion.

Whitespace

🔬 This is a nightly-only experimental API. (rustc_private)

Whitespace

Comment

🔬 This is a nightly-only experimental API. (rustc_private)

Comment

Shebang(Name)

🔬 This is a nightly-only experimental API. (rustc_private)

Eof

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl Token

fn is_like_gt(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token starts with '>'.

fn can_begin_expr(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token can appear at the start of an expression.

fn is_lit(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is any literal

fn is_ident(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is an identifier.

fn is_doc_comment(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is a documentation comment.

fn is_interpolated(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is interpolated.

fn is_path(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is an interpolated path.

fn is_lifetime(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is a lifetime.

fn is_mutability(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is either the mut or const keyword.

fn is_qpath_start(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_path_start(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_keyword(&self, kw: Keyword) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is a given keyword, kw.

fn is_path_segment_keyword(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_any_keyword(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is either a strict or reserved keyword.

fn is_strict_keyword(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is a strict keyword.

fn is_reserved_keyword(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if the token is a keyword reserved for possible future use.