Function syntax::print::pprust::print_crate

pub fn print_crate<'a>(
    cm: &'a CodeMap, 
    sess: &ParseSess, 
    krate: &Crate, 
    filename: String, 
    input: &mut Read, 
    out: Box<Write + 'a>, 
    ann: &'a PpAnn, 
    is_expanded: bool
) -> Result<()>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Requires you to pass an input filename and reader so that it can scan the input text for comments and literals to copy forward.