Struct syntax::ptr::P

pub struct P<T: ?Sized> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An owned smart pointer.

Methods

impl<T: 'static> P<T>

pub fn and_then<U, F>(self, f: F) -> U where
F: FnOnce(T) -> U,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Move out of the pointer. Intended for chaining transformations not covered by map.

pub fn unwrap(self) -> T

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Equivalent to and_then(|x| x)

pub fn map<F>(self, f: F) -> P<T> where
F: FnOnce(T) -> T,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Transform the inner value, consuming self and producing a new P<T>.

impl<T> P<[T]>

pub fn new() -> P<[T]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn from_vec(v: Vec<T>) -> P<[T]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn into_vec(self) -> Vec<T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?