Struct syntax::source_map::SourceFile

pub struct SourceFile {
    pub name: FileName,
    pub name_was_remapped: bool,
    pub unmapped_path: Option<FileName>,
    pub crate_of_origin: u32,
    pub src: Option<Rc<String>>,
    pub src_hash: u128,
    pub external_src: Lock<ExternalSource>,
    pub start_pos: BytePos,
    pub end_pos: BytePos,
    pub lines: Vec<BytePos>,
    pub multibyte_chars: Vec<MultiByteChar>,
    pub non_narrow_chars: Vec<NonNarrowChar>,
    pub name_hash: u128,
}

A single source in the SourceMap.

Fields
name: FileName

The name of the file that the source came from, source that doesn't originate from files has names between angle brackets by convention (e.g., <anon>).

name_was_remapped: bool

true if the name field above has been modified by --remap-path-prefix.

unmapped_path: Option<FileName>

The unmapped path of the file that the source came from. Set to None if the SourceFile was imported from an external crate.

crate_of_origin: u32

Indicates which crate this SourceFile was imported from.

src: Option<Rc<String>>

The complete source code.

src_hash: u128

The source code's hash.

external_src: Lock<ExternalSource>

The external source code (used for external crates, which will have a None value as self.src.

start_pos: BytePos

The start position of this source in the SourceMap.

end_pos: BytePos

The end position of this source in the SourceMap.

lines: Vec<BytePos>

Locations of lines beginnings in the source code.

multibyte_chars: Vec<MultiByteChar>

Locations of multi-byte characters in the source code.

non_narrow_chars: Vec<NonNarrowChar>

Width of characters that are not narrow in the source code.

name_hash: u128

A hash of the filename, used for speeding up hashing in incremental compilation.

Methods

impl SourceFile

pub fn new(
name: FileName,
name_was_remapped: bool,
unmapped_path: FileName,
src: String,
start_pos: BytePos
) -> SourceFile

pub fn line_begin_pos(&self, pos: BytePos) -> BytePos

Returns the BytePos of the beginning of the current line.

pub fn add_external_src<F>(&self, get_src: F) -> bool where
F: FnOnce() -> Option<String>,

Add externally loaded source. If the hash of the input doesn't match or no input is supplied via None, it is interpreted as an error and the corresponding enum variant is set. The return value signifies whether some kind of source is present.

pub fn get_line(&self, line_number: usize) -> Option<Cow<str>>

Gets a line from the list of pre-computed line-beginnings. The line number here is 0-based.

pub fn is_real_file(&self) -> bool

pub fn is_imported(&self) -> bool

pub fn byte_length(&self) -> u32

pub fn count_lines(&self) -> usize

pub fn lookup_line(&self, pos: BytePos) -> Option<usize>

Finds the line containing the given position. The return value is the index into the lines array of this SourceFile, not the 1-based line number. If the source_file is empty or the position is located before the first line, None is returned.

pub fn line_bounds(&self, line_index: usize) -> (BytePos, BytePos)

pub fn contains(&self, byte_pos: BytePos) -> bool