Struct syntax::source_map::SourceMap

pub struct SourceMap { /* fields omitted */ }

Methods

impl SourceMap

pub fn new(path_mapping: FilePathMapping) -> SourceMap

pub fn with_file_loader(
file_loader: Box<dyn FileLoader + Sync + Send>,
path_mapping: FilePathMapping
) -> SourceMap

pub fn path_mapping(&self) -> &FilePathMapping

pub fn file_exists(&self, path: &Path) -> bool

pub fn load_file(&self, path: &Path) -> Result<Lrc<SourceFile>>

pub fn files(&self) -> MappedLockGuard<Vec<Lrc<SourceFile>>>

pub fn source_file_by_stable_id(
&self,
stable_id: StableSourceFileId
) -> Option<Lrc<SourceFile>>

pub fn new_source_file(
&self,
filename: FileName,
src: String
) -> Lrc<SourceFile>

Creates a new source_file. If a file already exists in the source_map with the same id, that file is returned unmodified

pub fn new_imported_source_file(
&self,
filename: FileName,
name_was_remapped: bool,
crate_of_origin: u32,
src_hash: u128,
name_hash: u128,
source_len: usize,
file_local_lines: Vec<BytePos>,
file_local_multibyte_chars: Vec<MultiByteChar>,
file_local_non_narrow_chars: Vec<NonNarrowChar>
) -> Lrc<SourceFile>

Allocates a new SourceFile representing a source file from an external crate. The source code of such an "imported source_file" is not available, but we still know enough to generate accurate debuginfo location information for things inlined from other crates.

pub fn mk_substr_filename(&self, sp: Span) -> String

pub fn doctest_offset_line(&self, file: &FileName, orig: usize) -> usize

pub fn lookup_char_pos(&self, pos: BytePos) -> Loc

Lookup source information about a BytePos

pub fn lookup_line(
&self,
pos: BytePos
) -> Result<SourceFileAndLine, Lrc<SourceFile>>

pub fn lookup_char_pos_adj(&self, pos: BytePos) -> LocWithOpt

pub fn merge_spans(&self, sp_lhs: Span, sp_rhs: Span) -> Option<Span>

Returns Some(span), a union of the lhs and rhs span. The lhs must precede the rhs. If there are gaps between lhs and rhs, the resulting union will cross these gaps. For this to work, the spans have to be:

the ctxt of both spans much match

the lhs span needs to end on the same line the rhs span begins

the lhs span must start at or before the rhs span

pub fn span_to_string(&self, sp: Span) -> String

pub fn span_to_filename(&self, sp: Span) -> FileName

pub fn span_to_unmapped_path(&self, sp: Span) -> FileName

pub fn is_multiline(&self, sp: Span) -> bool

pub fn span_to_lines(&self, sp: Span) -> FileLinesResult

pub fn span_to_snippet(&self, sp: Span) -> Result<String, SpanSnippetError>

Returns the source snippet as String corresponding to the given Span

pub fn span_to_margin(&self, sp: Span) -> Option<usize>

pub fn span_to_prev_source(&self, sp: Span) -> Result<String, SpanSnippetError>

Returns the source snippet as String before the given Span

pub fn span_extend_to_prev_char(&self, sp: Span, c: char) -> Span

Extend the given Span to just after the previous occurrence of c. Return the same span if no character could be found or if an error occurred while retrieving the code snippet.

pub fn span_extend_to_prev_str(
&self,
sp: Span,
pat: &str,
accept_newlines: bool
) -> Span

Extend the given Span to just after the previous occurrence of pat when surrounded by whitespace. Return the same span if no character could be found or if an error occurred while retrieving the code snippet.

pub fn span_until_char(&self, sp: Span, c: char) -> Span

Given a Span, try to get a shorter span ending before the first occurrence of c char

pub fn span_through_char(&self, sp: Span, c: char) -> Span

Given a Span, try to get a shorter span ending just after the first occurrence of char c.

pub fn span_until_non_whitespace(&self, sp: Span) -> Span

Given a Span, get a new Span covering the first token and all its trailing whitespace or the original Span.

If sp points to "let mut x", then a span pointing at "let " will be returned.

pub fn span_until_whitespace(&self, sp: Span) -> Span

Given a Span, get a new Span covering the first token without its trailing whitespace or the original Span in case of error.

If sp points to "let mut x", then a span pointing at "let" will be returned.

pub fn span_take_while<P>(&self, sp: Span, predicate: P) -> Span where
P: for<'r> FnMut(&'r char) -> bool,

Given a Span, get a shorter one until predicate yields false.

pub fn def_span(&self, sp: Span) -> Span

pub fn start_point(&self, sp: Span) -> Span

Returns a new span representing just the start-point of this span

pub fn end_point(&self, sp: Span) -> Span

Returns a new span representing just the end-point of this span

pub fn next_point(&self, sp: Span) -> Span

Returns a new span representing the next character after the end-point of this span

pub fn get_source_file(&self, filename: &FileName) -> Option<Lrc<SourceFile>>

pub fn lookup_byte_offset(&self, bpos: BytePos) -> SourceFileAndBytePos

For a global BytePos compute the local offset within the containing SourceFile

pub fn bytepos_to_file_charpos(&self, bpos: BytePos) -> CharPos

Converts an absolute BytePos to a CharPos relative to the source_file.

pub fn lookup_source_file_idx(&self, pos: BytePos) -> usize

pub fn count_lines(&self) -> usize

pub fn generate_fn_name_span(&self, span: Span) -> Option<Span>

pub fn generate_local_type_param_snippet(
&self,
span: Span
) -> Option<(Span, String)>

Take the span of a type parameter in a function signature and try to generate a span for the function name (with generics) and a new snippet for this span with the pointed type parameter as a new local type parameter.

For instance:

// Given span
fn my_function(param: T)
//                    ^ Original span

// Result
fn my_function(param: T)
// ^^^^^^^^^^^ Generated span with snippet `my_function<T>`

Attention: The method used is very fragile since it essentially duplicates the work of the parser. If you need to use this function or something similar, please consider updating the source_map functions and this function to something more robust.