Struct syntax::source_map::Span

#[repr(packed)]

pub struct Span(_);

A compressed span. Contains either fields of SpanData inline if they are small, or index into span interner. The primary goal of Span is to be as small as possible and fit into other structures (that's why it uses packed as well). Decoding speed is the second priority. See SpanData for the info on span fields in decoded representation.

Methods

impl Span

pub fn new(lo: BytePos, hi: BytePos, ctxt: SyntaxContext) -> Span

pub fn data(self) -> SpanData

impl Span

pub fn lo(self) -> BytePos

pub fn with_lo(self, lo: BytePos) -> Span

pub fn hi(self) -> BytePos

pub fn with_hi(self, hi: BytePos) -> Span

pub fn ctxt(self) -> SyntaxContext

pub fn with_ctxt(self, ctxt: SyntaxContext) -> Span

pub fn is_dummy(self) -> bool

Returns true if this is a dummy span with any hygienic context.

pub fn shrink_to_lo(self) -> Span

Returns a new span representing an empty span at the beginning of this span

pub fn shrink_to_hi(self) -> Span

Returns a new span representing an empty span at the end of this span.

pub fn substitute_dummy(self, other: Span) -> Span

Returns self if self is not the dummy span, and other otherwise.

pub fn contains(self, other: Span) -> bool

Returns true if self fully encloses other.

pub fn overlaps(self, other: Span) -> bool

Returns true if self touches other.

pub fn source_equal(&self, other: &Span) -> bool

Returns true if the spans are equal with regards to the source text.

Use this instead of == when either span could be generated code, and you only care that they point to the same bytes of source text.

pub fn trim_start(self, other: Span) -> Option<Span>

Returns Some(span), where the start is trimmed by the end of other.

pub fn source_callsite(self) -> Span

Returns the source span -- this is either the supplied span, or the span for the macro callsite that expanded to it.

pub fn parent(self) -> Option<Span>

The Span for the tokens in the previous macro expansion from which self was generated, if any.

pub fn edition(self) -> Edition

Edition of the crate from which this span came.

pub fn rust_2015(&self) -> bool

pub fn rust_2018(&self) -> bool

pub fn source_callee(self) -> Option<ExpnInfo>

Returns the source callee.

Returns None if the supplied span has no expansion trace, else returns the ExpnInfo for the macro definition corresponding to the source callsite.

pub fn allows_unstable(&self, feature: &str) -> bool

Checks if a span is "internal" to a macro in which #[unstable] items can be used (that is, a macro marked with #[allow_internal_unstable]).

pub fn is_compiler_desugaring(&self, kind: CompilerDesugaringKind) -> bool

Checks if this span arises from a compiler desugaring of kind kind.

pub fn compiler_desugaring_kind(&self) -> Option<CompilerDesugaringKind>

Returns the compiler desugaring that created this span, or None if this span is not from a desugaring.

pub fn allows_unsafe(&self) -> bool

Checks if a span is "internal" to a macro in which unsafe can be used without triggering the unsafe_code lint

pub fn macro_backtrace(self) -> Vec<MacroBacktrace>

pub fn to(self, end: Span) -> Span

Returns a Span that would enclose both self and end.

pub fn between(self, end: Span) -> Span

Returns a Span between the end of self to the beginning of end.

pub fn until(self, end: Span) -> Span

Returns a Span between the beginning of self to the beginning of end.

pub fn from_inner_byte_pos(self, start: usize, end: usize) -> Span

pub fn apply_mark(self, mark: Mark) -> Span

pub fn remove_mark(&mut self) -> Mark

pub fn adjust(&mut self, expansion: Mark) -> Option<Mark>

pub fn glob_adjust(
&mut self,
expansion: Mark,
glob_ctxt: SyntaxContext
) -> Option<Option<Mark>>

pub fn reverse_glob_adjust(
&mut self,
expansion: Mark,
glob_ctxt: SyntaxContext
) -> Option<Option<Mark>>

pub fn modern(self) -> Span

pub fn modern_and_legacy(self) -> Span