Struct syntax::symbol::InternedString

pub struct InternedString { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Represents a string stored in the thread-local interner. Because the interner lives for the life of the thread, this can be safely treated as an immortal string, as long as it never crosses between threads.

FIXME(pcwalton): You must be careful about what you do in the destructors of objects stored in TLS, because they may run after the interner is destroyed. In particular, they must not access string contents. This can be fixed in the future by just leaking all strings until thread death somehow.