Struct syntax::symbol::SymbolIndex

pub struct SymbolIndex { /* fields omitted */ }

Methods

impl SymbolIndex

pub const MAX_AS_U32: u32

pub const MAX: SymbolIndex

pub fn from_usize(value: usize) -> SymbolIndex

pub fn from_u32(value: u32) -> SymbolIndex

pub const fn from_u32_const(value: u32) -> SymbolIndex

Hacky variant of from_u32 for use in constants. This version checks the "max" constraint by using an invalid array dereference.

pub const unsafe fn from_u32_unchecked(value: u32) -> SymbolIndex

pub fn index(self) -> usize

Extracts the value of this index as an integer.

pub fn as_u32(self) -> u32

Extracts the value of this index as a u32.

pub fn as_usize(self) -> usize

Extracts the value of this index as a usize.