syntax

Module syntax_editor

Syntax Tree editor

Inspired by Roslyn’s SyntaxEditor, but is temporarily built upon mutable syntax tree editing.

Structs

Position

Position describing where to insert elements

SyntaxAnnotation

SyntaxEdit

Represents a completed SyntaxEditor operation.

SyntaxEditor

SyntaxMapping

SyntaxMappingBuilder

Traits

Element

Utility trait to allow calling syntax editor functions with references or owned nodes. Do not use outside of this module.

Removable