Struct syntax::tokenstream::TokenSlice

pub struct TokenSlice(_);

Unstable (rustc_private)

TokenSlices are 'views' of TokenStream's; they fit the same role asstrs do forString`s. In general, most TokenStream manipulations will be refocusing their internal contents by taking a TokenSlice and then using indexing and the provided operators.

Methods

impl TokenSlice

fn is_empty(&self) -> bool

Unstable (rustc_private)

Indicates whether the TokenStream is empty.

fn len(&self) -> usize

Unstable (rustc_private)

Return the TokenSlice's length.

fn eq_unspanned(&self, other: &TokenSlice) -> bool

Unstable (rustc_private)

Check equality versus another TokenStream, ignoring span information.

fn covering_span(&self) -> Span

Unstable (rustc_private)

Compute a span that covers the entire TokenSlice (eg, one wide enough to include the entire slice). If the inputs share expansion identification, it is preserved. If they do not, it is discarded.

fn is_assignment(&self) -> bool

Unstable (rustc_private)

Indicates where the stream is of the form = <ts>, where <ts> is a continued TokenStream.

fn maybe_assignment(&self) -> Option<&TokenSlice>

Unstable (rustc_private)

Returns the RHS of an assigment.

fn is_delimited(&self) -> bool

Unstable (rustc_private)

Indicates where the stream is a single, delimited expression (e.g., (a,b,c) or {a,b,c}).

fn maybe_delimited(&self) -> Option<&TokenSlice>

Unstable (rustc_private)

Returns the inside of the delimited term as a new TokenStream.

fn maybe_comma_list(&self) -> Option<Vec<&TokenSlice>>

Unstable (rustc_private)

Returns a list of TokenSlices if the stream is a delimited list, breaking the stream on commas.

fn maybe_interpolated_nonterminal(&self) -> Option<Nonterminal>

Unstable (rustc_private)

Returns a Nonterminal if it is Interpolated.

fn is_ident(&self) -> bool

Unstable (rustc_private)

Indicates if the stream is exactly one identifier.

fn maybe_ident(&self) -> Option<Ident>

Unstable (rustc_private)

Returns an identifier

fn is_lit(&self) -> bool

Unstable (rustc_private)

Indicates if the stream is exactly one literal

fn maybe_lit(&self) -> Option<Lit>

Unstable (rustc_private)

Returns a literal

fn maybe_str(&self) -> Option<Lit>

Unstable (rustc_private)

Returns an AST string literal if the TokenStream is either a normal ('cooked') or raw string literal.

fn maybe_path_prefix(&self) -> Option<(Path, &TokenSlice)>

Unstable (rustc_private)

This operation extracts the path prefix , returning an AST path struct and the remainder of the stream (if it finds one). To be more specific, a tokenstream that has a valid, non-global path as a prefix (eg foo(bar, baz), foo::bar(bar), but not ::foo::bar(baz)) will yield the path and the remaining tokens (as a slice). The previous examples will yield Some((Path { segments = vec![foo], ... }, [(bar, baz)])), Some((Path { segments = vec![foo, bar] }, [(baz)])), and None, respectively.