Struct syntax::tokenstream::TokenStream

pub struct TokenStream {
    pub span: Span,
    pub tts: Vec<TokenTree>,
}

Unstable (rustc_private)

Token Streams

TokenStreams are a syntactic abstraction over TokenTrees. The goal is for procedural macros to work over TokenStreams instead of arbitrary syntax. For now, however, we are going to cut a few corners (i.e., use some of the AST structure) when we need to for backwards compatibility. TokenStreams are collections of TokenTrees that represent a syntactic structure. The struct itself shouldn't be directly manipulated; the internal structure is not stable, and may be changed at any time in the future. The operators will not, however (except for signatures, later on).

Fields
span: Span
Unstable (rustc_private)
tts: Vec<TokenTree>
Unstable (rustc_private)

Methods

impl TokenStream

TokenStream operators include basic destructuring, boolean operations, maybe_... operations, and maybe_..._prefix operations. Boolean operations are straightforward, indicating information about the structure of the stream. The maybe_... operations return Some<...> if the tokenstream contains the appropriate item.

Similarly, the maybe_..._prefix operations potentially return a partially-destructured stream as a pair where the first element is the expected item and the second is the remainder of the stream. As anb example,

maybe_path_prefix("a::b::c(a,b,c).foo()") -> (a::b::c, "(a,b,c).foo()")

fn from_tts(trees: Vec<TokenTree>) -> TokenStream

Unstable (rustc_private)

Convert a vector of TokenTrees into a TokenStream.

fn append_stream(self, ts2: &TokenSlice)

Unstable (rustc_private)

Copies all of the TokenTrees from the TokenSlice, appending them to the stream.

fn respan(self, span: Span) -> TokenStream

Unstable (rustc_private)

Manually change a TokenStream's span.

fn from_ast_lit_str(lit: Lit) -> Option<TokenStream>

Unstable (rustc_private)

Construct a TokenStream from an ast literal.

fn as_paren_delimited_stream(tts: Vec<TokenTree>) -> TokenStream

Unstable (rustc_private)

Convert a vector of TokenTrees into a parentheses-delimited TokenStream.

fn from_interned_string_as_ident(s: InternedString) -> TokenStream

Unstable (rustc_private)

Convert an interned string into a one-element TokenStream.