Struct syntax::util::interner::Interner

pub struct Interner { /* fields omitted */ }

Unstable (rustc_private)

Methods

impl Interner

When traits can extend traits, we should extend index to get []

fn new() -> Self

Unstable (rustc_private)

fn prefill(init: &[&str]) -> Self

Unstable (rustc_private)

fn intern<T: Borrow<str> + Into<String>>(&mut self, string: T) -> Name

Unstable (rustc_private)

fn gensym(&mut self, string: &str) -> Name

Unstable (rustc_private)

fn gensym_copy(&mut self, name: Name) -> Name

Unstable (rustc_private)

Create a gensym with the same name as an existing entry.

fn get(&self, name: Name) -> Rc<String>

Unstable (rustc_private)

fn find(&self, string: &str) -> Option<Name>

Unstable (rustc_private)