Trait syntax::util::move_map::MoveMap

pub trait MoveMap<T>: Sized {
    fn move_flat_map<F, I>(self, f: F) -> Self
    where
        F: FnMut(T) -> I,
        I: IntoIterator<Item = T>;

    fn move_map<F>(self, f: F) -> Self
    where
        F: FnMut(T) -> T,
    { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn move_flat_map<F, I>(self, f: F) -> Self where
F: FnMut(T) -> I,
I: IntoIterator<Item = T>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn move_map<F>(self, f: F) -> Self where
F: FnMut(T) -> T,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?