Module syntax::util::parser

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Enums

AssocOp 	 [
Experimental
]
Associative operator with precedence.

Fixity 	 [
Experimental
]

Constants

PREC_CLOSURE 	 [
Experimental
]

PREC_FORCE_PAREN 	 [
Experimental
]

PREC_JUMP 	 [
Experimental
]

PREC_PAREN 	 [
Experimental
]

PREC_POSTFIX 	 [
Experimental
]

PREC_PREFIX 	 [
Experimental
]

PREC_RANGE 	 [
Experimental
]

PREC_RESET 	 [
Experimental
]

Functions

contains_exterior_struct_lit 	 [
Experimental
]
Expressions that syntactically contain an "exterior" struct literal i.e. not surrounded by any parens or other delimiters, e.g. X { y: 1 }, X { y: 1 }.method(), foo == X { y: 1 } and X { y: 1 } == foo all do, but (X { y: 1 }) == foo does not.

expr_precedence 	 [
Experimental
]