Function syntax::visit::walk_expr

pub fn walk_expr<'a, V: Visitor<'a>>(visitor: &mut V, expression: &'a Expr)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?