Crate syntax

Unstable (rustc_private)

The Rust parser and macro expander.

Note

This API is completely unstable and subject to change.

Modules

abi 	 [Unstable]

ast 	 [Unstable]

attr 	 [Unstable]

codemap 	 [Unstable]
The CodeMap tracks all the source code used within a single crate, mapping from integer byte positions to the original source code location. Each bit of source parsed during crate parsing (typically files, in-memory strings, or various bits of macro expansion) cover a continuous range of bytes in the CodeMap and are represented by FileMaps. Byte positions are stored in spans and used pervasively in the compiler. They are absolute positions within the CodeMap, which upon request can be converted to line and column information, source code snippets, etc.

config 	 [Unstable]

diagnostics 	 [Unstable]

entry 	 [Unstable]

errors 	 [Unstable]

ext 	 [Unstable]

feature_gate 	 [Unstable]
Feature gating

fold 	 [Unstable]
A Folder represents an AST->AST fold; it accepts an AST piece, and returns a piece of the same type. So, for instance, macro expansion is a Folder that walks over an AST and produces another AST.

parse 	 [Unstable]
The main parser interface

print 	 [Unstable]

ptr 	 [Unstable]
The AST pointer

show_span 	 [Unstable]
Span debugger

std_inject 	 [Unstable]

str 	 [Unstable]

syntax 	 [Unstable]

test 	 [Unstable]

util 	 [Unstable]

visit 	 [Unstable]
AST walker. Each overridden visit method has full control over what happens with its node, it can do its own traversal of the node's children, call visit::walk_* to apply the default traversal algorithm, or prevent deeper traversal by doing nothing.

Macros

help! 	 [Unstable]

panictry! 	 [Unstable]

register_diagnostic! 	 [Unstable]

register_diagnostics! 	 [Unstable]

register_long_diagnostics! 	 [Unstable]

span_err! 	 [Unstable]

span_err_or_warn! 	 [Unstable]

span_fatal! 	 [Unstable]

span_help! 	 [Unstable]

span_note! 	 [Unstable]

span_warn! 	 [Unstable]

struct_span_err! 	 [Unstable]

struct_span_err_or_warn! 	 [Unstable]

struct_span_fatal! 	 [Unstable]

struct_span_warn! 	 [Unstable]

walk_list! 	 [Unstable]