Module syntax::ast

Modules

forget

fv

generic

source

subst

typed

untyped

working

Structs

Case

Cocase

Codata

Codef

Comatch

Const

Ctor

Data

Decls

Def

Dtor

Id

Impl

Match

Param

ParamInst

Instantiation of a previously declared parameter

Prg

SelfParam

Telescope

Wrapper type signifying the wrapped parameters have telescope semantics. I.e. each parameter binding in the parameter list is in scope for the following parameters.

TelescopeInst

Instantiation of a previously declared telescope

TypAbs

TypApp

Enums

Decl

Exp

Item

Type

Traits

Annotated

Fold

Folder

HasPhase

Map

Mapper

Out

Phase

Visit

Visitor

Functions

id

occurs_in

Type Definitions

Args

Params