Module syntax::ast

Structs

Assign

Case

Cocase

Codata

Codef

Comatch

Ctor

Data

Decls

Def

Dtor

Eqn

EqnParam

Info

Match

Param

Prg

Telescope

Wrapper type signifying the wrapped parameters have telescope semantics. I.e. each parameter binding in the parameter list is in scope for the following parameters.

TypAbs

TypApp

Enums

Decl

Exp

Traits

SubstTelescope

Substitutable

Substitution

Functions

occurs_in

Type Definitions

Args

EqnParams

Params