syntax

Module ast

Abstract Syntax Tree, layered on top of untyped SyntaxNodes

Modules

edit

This module contains functions for editing syntax trees. As the trees are immutable, all function here return a fresh copy of the tree, instead of doing an in-place modification.

edit_in_place

Structural editing for ast.

make

This module contains free-standing functions for creating AST fragments out of smaller pieces.

prec

Precedence representation.

syntax_factory

Builds upon crate::ast::make constructors to create ast fragments with optional syntax mappings.

Structs

Abi

AnyHasArgList

AnyHasAttrs

AnyHasDocComments

AnyHasGenericArgs

AnyHasGenericParams

AnyHasLoopBody

AnyHasModuleItem

AnyHasName

AnyHasTypeBounds

AnyHasVisibility

ArgList

ArrayExpr

ArrayType

AsmClobberAbi

AsmConst

AsmDirSpec

AsmExpr

AsmLabel

AsmOperandExpr

AsmOperandNamed

AsmOption

AsmOptions

AsmRegOperand

AsmRegSpec

AsmSym

AssocItemList

AssocTypeArg

AstChildren

An iterator over SyntaxNode children of a particular AST type.

Attr

AttrDocCommentIter

AwaitExpr

BecomeExpr

BinExpr

BlockExpr

BoxPat

BreakExpr

Byte

ByteString

CString

CallExpr

CastExpr

Char

ClosureExpr

Comment

CommentKind

Const

ConstArg

ConstBlockPat

ConstParam

ContinueExpr

DocCommentIter

DynTraitType

Enum

ExprStmt

ExternBlock

ExternCrate

ExternItemList

FieldExpr

FloatNumber

Fn

FnPtrType

ForBinder

ForExpr

ForType

FormatArgsArg

FormatArgsExpr

GenericArgList

GenericParamList

Ident

IdentPat

IfExpr

Impl

ImplTraitType

IndexExpr

InferType

IntNumber

ItemList

Label

LetElse

LetExpr

LetStmt

Lifetime

LifetimeArg

LifetimeParam

Literal

LiteralPat

LoopExpr

MacroCall

MacroDef

MacroExpr

MacroItems

MacroPat

MacroRules

MacroStmts

MacroType

MatchArm

MatchArmList

MatchExpr

MatchGuard

Meta

MethodCallExpr

Module

Name

NameRef

NeverType

OffsetOfExpr

OrPat

Param

ParamList

ParenExpr

ParenPat

ParenType

ParenthesizedArgList

Path

PathExpr

PathPat

PathSegment

PathType

PrefixExpr

PtrType

QuoteOffsets

RangeExpr

RangePat

RecordExpr

RecordExprField

RecordExprFieldList

RecordField

RecordFieldList

RecordPat

RecordPatField

RecordPatFieldList

RefExpr

RefPat

RefType

Rename

RestPat

RetType

ReturnExpr

ReturnTypeSyntax

SelfParam

SlicePat

SlicePatComponents

SliceType

SourceFile

Static

StmtList

String

Struct

TokenTree

Trait

TryExpr

TupleExpr

TupleField

TupleFieldList

TuplePat

TupleStructPat

TupleType

TypeAlias

TypeAnchor

TypeArg

TypeBound

TypeBoundList

TypeParam

UnderscoreExpr

Union

Use

UseBoundGenericArgs

UseTree

UseTreeList

Variant

VariantList

Visibility

WhereClause

WherePred

WhileExpr

Whitespace

WildcardPat

YeetExpr

YieldExpr

Enums

Adt

AnyString

ArithOp

ArrayExprKind

AsmOperand

AsmPiece

AssocItem

AttrKind

BinaryOp

BlockModifier

CallableExpr

CmpOp

CommentPlacement

CommentShape

ElseBranch

Expr

ExternItem

FieldKind

FieldList

GenericArg

GenericParam

Item

LiteralKind

LogicOp

Macro

NameLike

NameOrNameRef

Ordering

Pat

PathSegmentKind

Radix

RangeOp

SelfParamKind

Stmt

StructKind

Type

TypeBoundKind

TypeOrConstParam

UnaryOp

UseBoundGenericArg

VariantDef

VisibilityKind

Traits

AstNode

The main trait to go from untyped SyntaxNode to a typed ast. The conversion itself has zero runtime cost: ast and syntax nodes have exactly the same representation: a pointer to the tree root and a pointer to the node itself.

AstToken

Like AstNode, but wraps tokens rather than interior nodes.

HasArgList

HasAttrs

HasDocComments

HasGenericArgs

HasGenericParams

HasLoopBody

HasModuleItem

HasName

HasTypeBounds

HasVisibility

IsString

RangeItem

Trait to describe operations common to both RangeExpr and RangePat.