syntax::ast
Enum Expr

pub enum Expr {

    ArrayExpr(ArrayExpr),
    AsmExpr(AsmExpr),
    AwaitExpr(AwaitExpr),
    BecomeExpr(BecomeExpr),
    BinExpr(BinExpr),
    BlockExpr(BlockExpr),
    BreakExpr(BreakExpr),
    CallExpr(CallExpr),
    CastExpr(CastExpr),
    ClosureExpr(ClosureExpr),
    ContinueExpr(ContinueExpr),
    FieldExpr(FieldExpr),
    ForExpr(ForExpr),
    FormatArgsExpr(FormatArgsExpr),
    IfExpr(IfExpr),
    IndexExpr(IndexExpr),
    LetExpr(LetExpr),
    Literal(Literal),
    LoopExpr(LoopExpr),
    MacroExpr(MacroExpr),
    MatchExpr(MatchExpr),
    MethodCallExpr(MethodCallExpr),
    OffsetOfExpr(OffsetOfExpr),
    ParenExpr(ParenExpr),
    PathExpr(PathExpr),
    PrefixExpr(PrefixExpr),
    RangeExpr(RangeExpr),
    RecordExpr(RecordExpr),
    RefExpr(RefExpr),
    ReturnExpr(ReturnExpr),
    TryExpr(TryExpr),
    TupleExpr(TupleExpr),
    UnderscoreExpr(UnderscoreExpr),
    WhileExpr(WhileExpr),
    YeetExpr(YeetExpr),
    YieldExpr(YieldExpr),

}

Variants

ArrayExpr(ArrayExpr)

AsmExpr(AsmExpr)

AwaitExpr(AwaitExpr)

BecomeExpr(BecomeExpr)

BinExpr(BinExpr)

BlockExpr(BlockExpr)

BreakExpr(BreakExpr)

CallExpr(CallExpr)

CastExpr(CastExpr)

ClosureExpr(ClosureExpr)

ContinueExpr(ContinueExpr)

FieldExpr(FieldExpr)

ForExpr(ForExpr)

FormatArgsExpr(FormatArgsExpr)

IfExpr(IfExpr)

IndexExpr(IndexExpr)

LetExpr(LetExpr)

Literal(Literal)

LoopExpr(LoopExpr)

MacroExpr(MacroExpr)

MatchExpr(MatchExpr)

MethodCallExpr(MethodCallExpr)

OffsetOfExpr(OffsetOfExpr)

ParenExpr(ParenExpr)

PathExpr(PathExpr)

PrefixExpr(PrefixExpr)

RangeExpr(RangeExpr)

RecordExpr(RecordExpr)

RefExpr(RefExpr)

ReturnExpr(ReturnExpr)

TryExpr(TryExpr)

TupleExpr(TupleExpr)

UnderscoreExpr(UnderscoreExpr)

WhileExpr(WhileExpr)

YeetExpr(YeetExpr)

YieldExpr(YieldExpr)

Implementations

impl Expr

pub fn is_block_like(&self) -> bool

impl Expr

pub fn needs_parens_in(&self, parent: SyntaxNode) -> bool

Returns true if self would need to be wrapped in parentheses given that its parent is parent.

impl Expr

pub fn parse(text: &str, edition: Edition) -> Parse<Expr>

Parses an ast::Expr from text.

Note that if the parsed root node is not a valid expression, Parse::tree will panic. For example:

ast::Expr::parse("let fail = true;", Edition::CURRENT).tree();