Enum syntax::ast::ExprKind

pub enum ExprKind {
    Box(P<Expr>),
    ObsoleteInPlace(P<Expr>, P<Expr>),
    Array(Vec<P<Expr>>),
    Call(P<Expr>, Vec<P<Expr>>),
    MethodCall(PathSegment, Vec<P<Expr>>),
    Tup(Vec<P<Expr>>),
    Binary(BinOp, P<Expr>, P<Expr>),
    Unary(UnOp, P<Expr>),
    Lit(Lit),
    Cast(P<Expr>, P<Ty>),
    Type(P<Expr>, P<Ty>),
    If(P<Expr>, P<Block>, Option<P<Expr>>),
    IfLet(Vec<P<Pat>>, P<Expr>, P<Block>, Option<P<Expr>>),
    While(P<Expr>, P<Block>, Option<Label>),
    WhileLet(Vec<P<Pat>>, P<Expr>, P<Block>, Option<Label>),
    ForLoop(P<Pat>, P<Expr>, P<Block>, Option<Label>),
    Loop(P<Block>, Option<Label>),
    Match(P<Expr>, Vec<Arm>),
    Closure(CaptureBy, IsAsync, Movability, P<FnDecl>, P<Expr>, Span),
    Block(P<Block>, Option<Label>),
    Async(CaptureBy, NodeId, P<Block>),
    TryBlock(P<Block>),
    Assign(P<Expr>, P<Expr>),
    AssignOp(BinOp, P<Expr>, P<Expr>),
    Field(P<Expr>, Ident),
    Index(P<Expr>, P<Expr>),
    Range(Option<P<Expr>>, Option<P<Expr>>, RangeLimits),
    Path(Option<QSelf>, Path),
    AddrOf(Mutability, P<Expr>),
    Break(Option<Label>, Option<P<Expr>>),
    Continue(Option<Label>),
    Ret(Option<P<Expr>>),
    InlineAsm(P<InlineAsm>),
    Mac(Mac),
    Struct(Path, Vec<Field>, Option<P<Expr>>),
    Repeat(P<Expr>, AnonConst),
    Paren(P<Expr>),
    Try(P<Expr>),
    Yield(Option<P<Expr>>),
    Err,
}

Variants
Box(P<Expr>)

A box x expression.

ObsoleteInPlace(P<Expr>, P<Expr>)

First expr is the place; second expr is the value.

Array(Vec<P<Expr>>)

An array ([a, b, c, d])

Call(P<Expr>, Vec<P<Expr>>)

A function call

The first field resolves to the function itself, and the second field is the list of arguments. This also represents calling the constructor of tuple-like ADTs such as tuple structs and enum variants.

MethodCall(PathSegment, Vec<P<Expr>>)

A method call (x.foo::<'static, Bar, Baz>(a, b, c, d))

The PathSegment represents the method name and its generic arguments (within the angle brackets). The first element of the vector of an Expr is the expression that evaluates to the object on which the method is being called on (the receiver), and the remaining elements are the rest of the arguments. Thus, x.foo::<Bar, Baz>(a, b, c, d) is represented as ExprKind::MethodCall(PathSegment { foo, [Bar, Baz] }, [x, a, b, c, d]).

Tup(Vec<P<Expr>>)

A tuple (e.g., (a, b, c, d)).

Binary(BinOp, P<Expr>, P<Expr>)

A binary operation (e.g., a + b, a * b).

Unary(UnOp, P<Expr>)

A unary operation (e.g., !x, *x).

Lit(Lit)

A literal (e.g., 1, "foo").

Cast(P<Expr>, P<Ty>)

A cast (e.g., foo as f64).

Type(P<Expr>, P<Ty>)If(P<Expr>, P<Block>, Option<P<Expr>>)

An if block, with an optional else block.

if expr { block } else { expr }

IfLet(Vec<P<Pat>>, P<Expr>, P<Block>, Option<P<Expr>>)

An if let expression with an optional else block

if let pat = expr { block } else { expr }

This is desugared to a match expression.

While(P<Expr>, P<Block>, Option<Label>)

A while loop, with an optional label

'label: while expr { block }

WhileLet(Vec<P<Pat>>, P<Expr>, P<Block>, Option<Label>)

A while let loop, with an optional label.

'label: while let pat = expr { block }

This is desugared to a combination of loop and match expressions.

ForLoop(P<Pat>, P<Expr>, P<Block>, Option<Label>)

A for loop, with an optional label.

'label: for pat in expr { block }

This is desugared to a combination of loop and match expressions.

Loop(P<Block>, Option<Label>)

Conditionless loop (can be exited with break, continue, or return).

'label: loop { block }

Match(P<Expr>, Vec<Arm>)

A match block.

Closure(CaptureBy, IsAsync, Movability, P<FnDecl>, P<Expr>, Span)

A closure (e.g., move |a, b, c| a + b + c).

The final span is the span of the argument block |...|.

Block(P<Block>, Option<Label>)

A block ('label: { ... }).

Async(CaptureBy, NodeId, P<Block>)

An async block (async move { ... }).

The NodeId is the NodeId for the closure that results from desugaring an async block, just like the NodeId field in the IsAsync enum. This is necessary in order to create a def for the closure which can be used as a parent of any child defs. Defs created during lowering cannot be made the parent of any other preexisting defs.

TryBlock(P<Block>)

A try block (try { ... }).

Assign(P<Expr>, P<Expr>)

An assignment (a = foo()).

AssignOp(BinOp, P<Expr>, P<Expr>)

An assignment with an operator.

E.g., a += 1.

Field(P<Expr>, Ident)

Access of a named (e.g., obj.foo) or unnamed (e.g., obj.0) struct field.

Index(P<Expr>, P<Expr>)

An indexing operation (e.g., foo[2]).

Range(Option<P<Expr>>, Option<P<Expr>>, RangeLimits)

A range (e.g., 1..2, 1.., ..2, 1...2, 1..., ...2).

Path(Option<QSelf>, Path)

Variable reference, possibly containing :: and/or type parameters (e.g., foo::bar::<baz>).

Optionally "qualified" (e.g., <Vec<T> as SomeTrait>::SomeType).

AddrOf(Mutability, P<Expr>)

A referencing operation (&a or &mut a).

Break(Option<Label>, Option<P<Expr>>)

A break, with an optional label to break, and an optional expression.

Continue(Option<Label>)

A continue, with an optional label.

Ret(Option<P<Expr>>)

A return, with an optional value to be returned.

InlineAsm(P<InlineAsm>)

Output of the asm!() macro.

Mac(Mac)

A macro invocation; pre-expansion.

Struct(Path, Vec<Field>, Option<P<Expr>>)

A struct literal expression.

E.g., Foo {x: 1, y: 2}, or Foo {x: 1, .. base}, where base is the Option<Expr>.

Repeat(P<Expr>, AnonConst)

An array literal constructed from one repeated element.

E.g., [1; 5]. The expression is the element to be repeated; the constant is the number of times to repeat it.

Paren(P<Expr>)

No-op: used solely so we can pretty-print faithfully.

Try(P<Expr>)

A try expression (expr?).

Yield(Option<P<Expr>>)

A yield, with an optional value to be yielded.

Err

Placeholder for an expression that wasn't syntactically well formed in some way.