syntax::ast::syntax_factory

Struct SyntaxFactory

pub struct SyntaxFactory { /* private fields */ }

Implementations

impl SyntaxFactory

pub fn name(&self, name: &str) -> Name

pub fn name_ref(&self, name: &str) -> NameRef

pub fn lifetime(&self, text: &str) -> Lifetime

pub fn ty(&self, text: &str) -> Type

pub fn ty_infer(&self) -> InferType

pub fn ty_path(&self, path: Path) -> PathType

pub fn type_param(&self, name: Name, bounds: Option<TypeBoundList>) -> TypeParam

pub fn path_segment(&self, name_ref: NameRef) -> PathSegment

pub fn path_segment_generics( &self, name_ref: NameRef, generic_arg_list: GenericArgList, ) -> PathSegment

pub fn use_( &self, attrs: impl IntoIterator<Item = Attr>, visibility: Option<Visibility>, use_tree: UseTree, ) -> Use

pub fn use_tree( &self, path: Path, use_tree_list: Option<UseTreeList>, alias: Option<Rename>, add_star: bool, ) -> UseTree

pub fn path_unqualified(&self, segment: PathSegment) -> Path

pub fn path_from_segments( &self, segments: impl IntoIterator<Item = PathSegment>, is_abs: bool, ) -> Path

pub fn ident_pat(&self, ref_: bool, mut_: bool, name: Name) -> IdentPat

pub fn wildcard_pat(&self) -> WildcardPat

pub fn literal_pat(&self, text: &str) -> LiteralPat

pub fn slice_pat(&self, pats: impl IntoIterator<Item = Pat>) -> SlicePat

pub fn tuple_pat(&self, pats: impl IntoIterator<Item = Pat>) -> TuplePat

pub fn tuple_struct_pat( &self, path: Path, fields: impl IntoIterator<Item = Pat>, ) -> TupleStructPat

pub fn record_pat_with_fields( &self, path: Path, fields: RecordPatFieldList, ) -> RecordPat

pub fn record_pat_field_list( &self, fields: impl IntoIterator<Item = RecordPatField>, rest_pat: Option<RestPat>, ) -> RecordPatFieldList

pub fn record_pat_field(&self, name_ref: NameRef, pat: Pat) -> RecordPatField

pub fn record_pat_field_shorthand(&self, pat: Pat) -> RecordPatField

pub fn path_pat(&self, path: Path) -> Pat

pub fn rest_pat(&self) -> RestPat

pub fn or_pat( &self, pats: impl IntoIterator<Item = Pat>, leading_pipe: bool, ) -> OrPat

pub fn box_pat(&self, pat: Pat) -> BoxPat

pub fn paren_pat(&self, pat: Pat) -> ParenPat

pub fn range_pat(&self, start: Option<Pat>, end: Option<Pat>) -> RangePat

pub fn ref_pat(&self, pat: Pat) -> RefPat

pub fn block_expr( &self, statements: impl IntoIterator<Item = Stmt>, tail_expr: Option<Expr>, ) -> BlockExpr

pub fn expr_empty_block(&self) -> BlockExpr

pub fn expr_paren(&self, expr: Expr) -> ParenExpr

pub fn expr_tuple(&self, fields: impl IntoIterator<Item = Expr>) -> TupleExpr

pub fn expr_assignment(&self, lhs: Expr, rhs: Expr) -> BinExpr

pub fn expr_bin(&self, lhs: Expr, op: BinaryOp, rhs: Expr) -> BinExpr

pub fn expr_literal(&self, text: &str) -> Literal

pub fn expr_path(&self, path: Path) -> Expr

pub fn expr_prefix(&self, op: SyntaxKind, expr: Expr) -> PrefixExpr

pub fn expr_call(&self, expr: Expr, arg_list: ArgList) -> CallExpr

pub fn expr_method_call( &self, receiver: Expr, method: NameRef, arg_list: ArgList, ) -> MethodCallExpr

pub fn arg_list(&self, args: impl IntoIterator<Item = Expr>) -> ArgList

pub fn expr_ref(&self, expr: Expr, exclusive: bool) -> Expr

pub fn expr_closure( &self, pats: impl IntoIterator<Item = Param>, expr: Expr, ) -> ClosureExpr

pub fn expr_return(&self, expr: Option<Expr>) -> ReturnExpr

pub fn expr_underscore(&self) -> UnderscoreExpr

pub fn expr_if( &self, condition: Expr, then_branch: BlockExpr, else_branch: Option<ElseBranch>, ) -> IfExpr

pub fn expr_loop(&self, body: BlockExpr) -> LoopExpr

pub fn expr_while_loop(&self, condition: Expr, body: BlockExpr) -> WhileExpr

pub fn expr_let(&self, pattern: Pat, expr: Expr) -> LetExpr

pub fn expr_stmt(&self, expr: Expr) -> ExprStmt

pub fn expr_match(&self, expr: Expr, match_arm_list: MatchArmList) -> MatchExpr

pub fn expr_macro(&self, path: Path, tt: TokenTree) -> MacroExpr

pub fn match_arm( &self, pat: Pat, guard: Option<MatchGuard>, expr: Expr, ) -> MatchArm

pub fn match_guard(&self, condition: Expr) -> MatchGuard

pub fn match_arm_list( &self, match_arms: impl IntoIterator<Item = MatchArm>, ) -> MatchArmList

pub fn let_stmt( &self, pattern: Pat, ty: Option<Type>, initializer: Option<Expr>, ) -> LetStmt

pub fn let_else_stmt( &self, pattern: Pat, ty: Option<Type>, initializer: Expr, diverging: BlockExpr, ) -> LetStmt

pub fn type_arg(&self, ty: Type) -> TypeArg

pub fn lifetime_arg(&self, lifetime: Lifetime) -> LifetimeArg

pub fn item_const( &self, attrs: impl IntoIterator<Item = Attr>, visibility: Option<Visibility>, name: Name, ty: Type, expr: Expr, ) -> Const

pub fn item_static( &self, visibility: Option<Visibility>, is_unsafe: bool, is_mut: bool, name: Name, ty: Type, expr: Option<Expr>, ) -> Static

pub fn param(&self, pat: Pat, ty: Type) -> Param

pub fn generic_arg_list( &self, generic_args: impl IntoIterator<Item = GenericArg>, is_turbo: bool, ) -> GenericArgList

pub fn record_expr(&self, path: Path, fields: RecordExprFieldList) -> RecordExpr

pub fn record_expr_field( &self, name: NameRef, expr: Option<Expr>, ) -> RecordExprField

pub fn record_field_list( &self, fields: impl IntoIterator<Item = RecordField>, ) -> RecordFieldList

pub fn record_field( &self, visibility: Option<Visibility>, name: Name, ty: Type, ) -> RecordField

pub fn tuple_field_list( &self, fields: impl IntoIterator<Item = TupleField>, ) -> TupleFieldList

pub fn tuple_field( &self, visibility: Option<Visibility>, ty: Type, ) -> TupleField

pub fn item_enum( &self, attrs: impl IntoIterator<Item = Attr>, visibility: Option<Visibility>, name: Name, generic_param_list: Option<GenericParamList>, where_clause: Option<WhereClause>, variant_list: VariantList, ) -> Enum

pub fn variant_list( &self, variants: impl IntoIterator<Item = Variant>, ) -> VariantList

pub fn variant( &self, visibility: Option<Visibility>, name: Name, field_list: Option<FieldList>, discriminant: Option<Expr>, ) -> Variant

pub fn fn_( &self, attrs: impl IntoIterator<Item = Attr>, visibility: Option<Visibility>, fn_name: Name, type_params: Option<GenericParamList>, where_clause: Option<WhereClause>, params: ParamList, body: BlockExpr, ret_type: Option<RetType>, is_async: bool, is_const: bool, is_unsafe: bool, is_gen: bool, ) -> Fn

pub fn attr_outer(&self, meta: Meta) -> Attr

pub fn attr_inner(&self, meta: Meta) -> Attr

pub fn meta_token_tree(&self, path: Path, tt: TokenTree) -> Meta

pub fn token_tree( &self, delimiter: SyntaxKind, tt: impl IntoIterator<Item = NodeOrToken<TokenTree, SyntaxToken>>, ) -> TokenTree

pub fn token(&self, kind: SyntaxKind) -> SyntaxToken

pub fn whitespace(&self, text: &str) -> SyntaxToken

pub fn ident(&self, text: &str) -> SyntaxToken

impl SyntaxFactory

pub fn ident_path(&self, ident: &str) -> Path

pub fn expr_unit(&self) -> Expr

pub fn ty_option(&self, t: Type) -> PathType

pub fn ty_result(&self, t: Type, e: Type) -> PathType

impl SyntaxFactory

pub fn with_mappings() -> Self

Creates a new SyntaxFactory, generating mappings between input nodes and generated nodes.

pub fn without_mappings() -> Self

Creates a SyntaxFactory without generating mappings.

pub fn finish_with_mappings(self) -> SyntaxMapping

Gets all of the tracked syntax mappings, if any.

pub fn take(&self) -> SyntaxMapping

Take all of the tracked syntax mappings, leaving SyntaxMapping::default() in its place, if any.