Trait syntax::ast::HasAttrs

pub trait HasAttrs: AstNode {
    // Provided methods
    fn attrs(&self) -> AstChildren<Attr>  { ... }
    fn has_atom_attr(&self, atom: &str) -> bool { ... }
    fn attrs_including_inner(self) -> impl Iterator<Item = Attr>
       where Self: Sized { ... }
}

Provided Methods

fn attrs(&self) -> AstChildren<Attr>

fn has_atom_attr(&self, atom: &str) -> bool

fn attrs_including_inner(self) -> impl Iterator<Item = Attr>
where Self: Sized,

Returns all attributes of this node, including inner attributes that may not be directly under this node but under a child.