Module syntax::cst

Structs

Case

Cocase

Codata

Codef

Comatch

Ctor

Data

Def

Dtor

Impl

Info

Match

Param

ParamInst

Instantiation of a previously declared parameter

Prg

Telescope

Wrapper type signifying the wrapped parameters have telescope semantics. I.e. each parameter binding in the parameter list is in scope for the following parameters. This influences the lowering semantic.

TelescopeInst

Instantiation of a previously declared telescope

TypApp

Enums

DefDecl

Exp

Item

TypDecl

Type Definitions

Args

Params