Trait syntax::ctx::Bind

pub trait Bind {
    type Elem;

    fn bind_fold<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
        &mut self,
        iter: I,
        acc: O1,
        f_acc: F1,
        f_inner: F2
    ) -> O2
    where
        T: AsElement<Self::Elem>,
        F1: Fn(&mut Self, O1, T) -> O1,
        F2: FnOnce(&mut Self, O1) -> O2;
    fn bind_fold2<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
        &mut self,
        iter: I,
        acc: O1,
        f_acc: F1,
        f_inner: F2
    ) -> O2
    where
        F1: Fn(&mut Self, O1, T) -> BindElem<Self::Elem, O1>,
        F2: FnOnce(&mut Self, O1) -> O2;
    fn bind_fold_failable<T, I: Iterator<Item = T>, O1, O2, F1, F2, E>(
        &mut self,
        iter: I,
        acc: O1,
        f_acc: F1,
        f_inner: F2
    ) -> Result<O2, E>
    where
        F1: Fn(&mut Self, O1, T) -> Result<BindElem<Self::Elem, O1>, E>,
        F2: FnOnce(&mut Self, O1) -> O2;

    fn bind_single<T, O, F>(&mut self, elem: T, f: F) -> O
    where
        T: AsElement<Self::Elem>,
        F: Fn(&mut Self) -> O,
    { ... }
    fn bind_iter<T, I, O, F>(&mut self, iter: I, f: F) -> O
    where
        T: AsElement<Self::Elem>,
        I: Iterator<Item = T>,
        F: FnOnce(&mut Self) -> O,
    { ... }
}

Required Associated Types

type Elem

Required Methods

fn bind_fold<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> O2where
T: AsElement<Self::Elem>,
F1: Fn(&mut Self, O1, T) -> O1,
F2: FnOnce(&mut Self, O1) -> O2,

Bind an iterator iter of elements

Fold the iterator and consume the result under the inner context with all binders in scope.

This is used for checking telescopes.

iter - An iterator of binders

acc - Accumulator for folding the iterator

f_acc - Accumulator function run for each binder

f_inner - Inner function computing the final result under the context of all binders

fn bind_fold2<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> O2where
F1: Fn(&mut Self, O1, T) -> BindElem<Self::Elem, O1>,
F2: FnOnce(&mut Self, O1) -> O2,

fn bind_fold_failable<T, I: Iterator<Item = T>, O1, O2, F1, F2, E>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> Result<O2, E>where
F1: Fn(&mut Self, O1, T) -> Result<BindElem<Self::Elem, O1>, E>,
F2: FnOnce(&mut Self, O1) -> O2,

Provided Methods

fn bind_single<T, O, F>(&mut self, elem: T, f: F) -> Owhere
T: AsElement<Self::Elem>,
F: Fn(&mut Self) -> O,

Bind a single element

fn bind_iter<T, I, O, F>(&mut self, iter: I, f: F) -> Owhere
T: AsElement<Self::Elem>,
I: Iterator<Item = T>,
F: FnOnce(&mut Self) -> O,

Bind an iterator iter of binders