Module syntax::diagnosticUnstable

Reexports

pub use self::Level::*;

pub use self::RenderSpan::*;

pub use self::ColorConfig::*;

Structs

EmitterWriter

ExplicitBug
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError
This structure is used to signify that a task has panicked with a fatal error from the diagnostics. You can use this with the Any trait to figure out how a rustc task died (if so desired).

Handler
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

SpanHandler
A span-handler is like a handler but also accepts span information for source-location reporting.

Enums

ColorConfig

Level

RenderSpan

Traits

Emitter

Functions

default_handler

expect

mk_handler

mk_span_handler