Enum syntax::ext::base::SyntaxExtension

pub enum SyntaxExtension {
    MultiDecorator(Box<MultiItemDecorator>),
    MultiModifier(Box<MultiItemModifier>),
    ProcMacro(Box<ProcMacro>),
    AttrProcMacro(Box<AttrProcMacro>),
    NormalTT(Box<TTMacroExpander>, Option<Span>, bool),
    IdentTT(Box<IdentMacroExpander>, Option<Span>, bool),
    CustomDerive(Box<MultiItemModifier>),
}

Unstable (rustc_private)

An enum representing the different kinds of syntax extensions.

Variants
MultiDecorator(Box<MultiItemDecorator>)
Unstable (rustc_private)

A syntax extension that is attached to an item and creates new items based upon it.

#[derive(...)] is a MultiItemDecorator.

Prefer ProcMacro or MultiModifier since they are more flexible.

MultiModifier(Box<MultiItemModifier>)
Unstable (rustc_private)

A syntax extension that is attached to an item and modifies it in-place. Also allows decoration, i.e., creating new items.

ProcMacro(Box<ProcMacro>)
Unstable (rustc_private)

A function-like procedural macro. TokenStream -> TokenStream.

AttrProcMacro(Box<AttrProcMacro>)
Unstable (rustc_private)

An attribute-like procedural macro. TokenStream, TokenStream -> TokenStream. The first TokenSteam is the attribute, the second is the annotated item. Allows modification of the input items and adding new items, similar to MultiModifier, but uses TokenStreams, rather than AST nodes.

NormalTT(Box<TTMacroExpander>, Option<Span>, bool)
Unstable (rustc_private)

A normal, function-like syntax extension.

bytes! is a NormalTT.

The bool dictates whether the contents of the macro can directly use #[unstable] things (true == yes).

IdentTT(Box<IdentMacroExpander>, Option<Span>, bool)
Unstable (rustc_private)

A function-like syntax extension that has an extra ident before the block.

CustomDerive(Box<MultiItemModifier>)
Unstable (rustc_private)