Enum syntax::ext::base::SyntaxExtension

pub enum SyntaxExtension {
    MultiDecorator(Box<MultiItemDecorator>),
    MultiModifier(Box<MultiItemModifier>),
    ProcMacro(Box<ProcMacro>),
    AttrProcMacro(Box<AttrProcMacro>),
    NormalTT {
        expander: Box<TTMacroExpander>,
        def_info: Option<(NodeId, Span)>,
        allow_internal_unstable: bool,
        allow_internal_unsafe: bool,
    },
    IdentTT(Box<IdentMacroExpander>, Option<Span>, bool),
    ProcMacroDerive(Box<MultiItemModifier>, Vec<Symbol>),
    BuiltinDerive(BuiltinDeriveFn),
    DeclMacro(Box<TTMacroExpander>, Option<(NodeId, Span)>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An enum representing the different kinds of syntax extensions.

Variants
MultiDecorator(Box<MultiItemDecorator>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A syntax extension that is attached to an item and creates new items based upon it.

#[derive(...)] is a MultiItemDecorator.

Prefer ProcMacro or MultiModifier since they are more flexible.

MultiModifier(Box<MultiItemModifier>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A syntax extension that is attached to an item and modifies it in-place. Also allows decoration, i.e., creating new items.

ProcMacro(Box<ProcMacro>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A function-like procedural macro. TokenStream -> TokenStream.

AttrProcMacro(Box<AttrProcMacro>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An attribute-like procedural macro. TokenStream, TokenStream -> TokenStream. The first TokenSteam is the attribute, the second is the annotated item. Allows modification of the input items and adding new items, similar to MultiModifier, but uses TokenStreams, rather than AST nodes.

NormalTT

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A normal, function-like syntax extension.

bytes! is a NormalTT.

Fields of NormalTT

expander: Box<TTMacroExpander>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

def_info: Option<(NodeId, Span)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

allow_internal_unstable: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Whether the contents of the macro can directly use #[unstable] things (true == yes).

allow_internal_unsafe: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Whether the contents of the macro can use unsafe without triggering the unsafe_code lint.

IdentTT(Box<IdentMacroExpander>, Option<Span>, bool)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A function-like syntax extension that has an extra ident before the block.

ProcMacroDerive(Box<MultiItemModifier>, Vec<Symbol>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An attribute-like procedural macro. TokenStream -> TokenStream. The input is the annotated item. Allows generating code to implement a Trait for a given struct or enum item.

BuiltinDerive(BuiltinDeriveFn)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An attribute-like procedural macro that derives a builtin trait.

DeclMacro(Box<TTMacroExpander>, Option<(NodeId, Span)>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A declarative macro, e.g. macro m() {}.

The second element is the definition site span.

Methods

impl SyntaxExtension

pub fn kind(&self) -> MacroKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Return which kind of macro calls this syntax extension.

pub fn is_modern(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?