Struct syntax::ext::base::DummyResult

pub struct DummyResult { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fill-in macro expansion result, to allow compilation to continue after hitting errors.

Methods

impl DummyResult

pub fn any(sp: Span) -> Box<MacResult + 'static>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a default MacResult that can be anything.

Use this as a return value after hitting any errors and calling span_err.

pub fn expr(sp: Span) -> Box<MacResult + 'static>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a default MacResult that can only be an expression.

Use this for macros that must expand to an expression, so even if an error is encountered internally, the user will receive an error that they also used it in the wrong place.

pub fn raw_expr(sp: Span) -> P<Expr>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A plain dummy expression.

pub fn raw_pat(sp: Span) -> Pat

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A plain dummy pattern.

pub fn raw_ty(sp: Span) -> P<Ty>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?