Struct syntax::ext::base::SyntaxEnv

pub struct SyntaxEnv {
    // some fields omitted
}

Unstable

In order to have some notion of scoping for macros, we want to implement the notion of a transformation environment.

This environment maps Names to SyntaxExtensions.

Methods

impl SyntaxEnv

fn push_frame(&mut self)

Unstable

fn pop_frame(&mut self)

Unstable

fn find(&self, k: &Name) -> Option<Rc<SyntaxExtension>>

Unstable

fn insert(&mut self, k: Name, v: SyntaxExtension)

Unstable

fn info<'a>(&'a mut self) -> &'a mut BlockInfo

Unstable