Struct syntax::ext::base::SyntaxEnv

pub struct SyntaxEnv {
    pub names: HashSet<Name>,
    // some fields omitted
}

Unstable (rustc_private)

In order to have some notion of scoping for macros, we want to implement the notion of a transformation environment.

This environment maps Names to SyntaxExtensions.

Fields

names
Unstable (rustc_private)

All bang-style macro/extension names encountered so far; to be used for diagnostics in resolve

Methods

impl SyntaxEnv

fn push_frame(&mut self)

Unstable (rustc_private)

fn pop_frame(&mut self)

Unstable (rustc_private)

fn find(&self, k: Name) -> Option<Rc<SyntaxExtension>>

Unstable (rustc_private)

fn insert(&mut self, k: Name, v: SyntaxExtension)

Unstable (rustc_private)

fn info(&mut self) -> &mut BlockInfo

Unstable (rustc_private)