Trait syntax::ext::base::MacResult

pub trait MacResult {
    fn make_expr(self: Box<Self>) -> Option<P<Expr>> { ... }
    fn make_items(self: Box<Self>) -> Option<SmallVector<P<Item>>> { ... }
    fn make_impl_items(self: Box<Self>) -> Option<SmallVector<ImplItem>> { ... }
    fn make_pat(self: Box<Self>) -> Option<P<Pat>> { ... }
    fn make_stmts(self: Box<Self>) -> Option<SmallVector<Stmt>> { ... }
    fn make_ty(self: Box<Self>) -> Option<P<Ty>> { ... }
}

Unstable (rustc_private)

The result of a macro expansion. The return values of the various methods are spliced into the AST at the callsite of the macro.

Provided Methods

fn make_expr(self: Box<Self>) -> Option<P<Expr>>

Unstable (rustc_private)

Create an expression.

fn make_items(self: Box<Self>) -> Option<SmallVector<P<Item>>>

Unstable (rustc_private)

Create zero or more items.

fn make_impl_items(self: Box<Self>) -> Option<SmallVector<ImplItem>>

Unstable (rustc_private)

Create zero or more impl items.

fn make_pat(self: Box<Self>) -> Option<P<Pat>>

Unstable (rustc_private)

Create a pattern.

fn make_stmts(self: Box<Self>) -> Option<SmallVector<Stmt>>

Unstable (rustc_private)

Create zero or more statements.

By default this attempts to create an expression statement, returning None if that fails.

fn make_ty(self: Box<Self>) -> Option<P<Ty>>

Unstable (rustc_private)