Struct syntax::ext::deriving::generic::FieldInfoUnstable

pub struct FieldInfo {
    pub span: Span,
    pub name: Option<Ident>,
    pub self_: P<Expr>,
    pub other: Vec<P<Expr>>,
}

Summary of the relevant parts of a struct/enum field.

Fields

span

name

None for tuple structs/normal enum variants, Some for normal structs/struct enum variants.

self_

The expression corresponding to this field of self (specifically, a reference to it).

other

The expressions corresponding to references to this field in the other Self arguments.