Module syntax::parse::parser

Unstable (rustc_private)

Structs

ModulePath 	 [Unstable]
Information about the path to a module.

ModulePathError 	 [Unstable]

ModulePathSuccess 	 [Unstable]

Parser 	 [Unstable]

Restrictions 	 [Unstable]

Enums

BoundParsingMode 	 [Unstable]
How to parse a bound, whether to allow bound modifiers such as ?.

LhsExpr 	 [Unstable]

PathStyle 	 [Unstable]
How to parse a path. There are three different kinds of paths, all of which are parsed somewhat differently.

SemiColonMode 	 [Unstable]

TokenType 	 [Unstable]