Module syntax::parse::parserUnstable

Reexports

pub use self::PathParsingMode::*;

Structs

Parser

Restrictions

Enums

BoundParsingMode
How to parse a bound, whether to allow bound modifiers such as ?.

PathParsingMode
How to parse a path. There are four different kinds of paths, all of which are parsed somewhat differently.

TokenType

Constants

RESTRICTION_NO_BAR_OP

RESTRICTION_NO_STRUCT_LITERAL

RESTRICTION_STMT_EXPR

UNRESTRICTED