Enum syntax::parse::token::Token

pub enum Token {
    Eq,
    Lt,
    Le,
    EqEq,
    Ne,
    Ge,
    Gt,
    AndAnd,
    OrOr,
    Not,
    Tilde,
    BinOp(BinOpToken),
    BinOpEq(BinOpToken),
    At,
    Dot,
    DotDot,
    DotDotDot,
    DotDotEq,
    Comma,
    Semi,
    Colon,
    ModSep,
    RArrow,
    LArrow,
    FatArrow,
    Pound,
    Dollar,
    Question,
    SingleQuote,
    OpenDelim(DelimToken),
    CloseDelim(DelimToken),
    Literal(Lit, Option<Name>),
    Ident(Ident, bool),
    Lifetime(Ident),
    Interpolated(Lrc<Nonterminal>),
    DocComment(Name),
    Whitespace,
    Comment,
    Shebang(Name),
    Eof,
}

Variants
EqLtLeEqEqNeGeGtAndAndOrOrNotTildeBinOp(BinOpToken)BinOpEq(BinOpToken)AtDotDotDotDotDotDotDotDotEqCommaSemiColonModSepRArrowLArrowFatArrowPoundDollarQuestionSingleQuote

Used by proc macros for representing lifetimes, not generated by lexer right now.

OpenDelim(DelimToken)

An opening delimiter (e.g., {).

CloseDelim(DelimToken)

A closing delimiter (e.g., }).

Literal(Lit, Option<Name>)Ident(Ident, bool)Lifetime(Ident)Interpolated(Lrc<Nonterminal>)DocComment(Name)

A doc comment.

Whitespace

Whitespace.

Comment

A comment.

Shebang(Name)Eof
Methods

impl Token

pub fn from_ast_ident(ident: Ident) -> Token

Recovers a Token from an ast::Ident. This creates a raw identifier if necessary.

pub fn can_begin_const_arg(&self) -> bool

Returns true if the token can appear at the start of a const param.

pub fn ident(&self) -> Option<(Ident, bool)>

Returns an identifier if this token is an identifier.

pub fn lifetime(&self) -> Option<Ident>

Returns a lifetime identifier if this token is a lifetime.

pub fn is_ident(&self) -> bool

Returns true if the token is an identifier.

pub fn is_keyword(&self, kw: Keyword) -> bool

Returns true if the token is a given keyword, kw.

pub fn is_path_segment_keyword(&self) -> bool

pub fn is_special_ident(&self) -> bool

pub fn is_reserved_ident(&self) -> bool

Returns true if the token is either a special identifier or a keyword.