syntax
Module ted

Primitive tree editor, ed for trees.

The _raw-suffixed functions insert elements as is, unsuffixed versions fix up elements around the edges.

Structs

Position

Traits

Element

Utility trait to allow calling ted functions with references or owned nodes. Do not use outside of this module.

Functions

append_child

append_child_raw

insert

insert_all

insert_all_raw

insert_raw

prepend_child

remove

remove_all

remove_all_iter

replace

replace_all

replace_with_many