Struct syntax::tokenstream::TokenStream

pub struct TokenStream { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

Token Streams

A TokenStream is an abstract sequence of tokens, organized into TokenTrees. The goal is for procedural macros to work with TokenStreams and TokenTrees instead of a representation of the abstract syntax tree. Today's TokenTrees can still contain AST via Token::Interpolated for back-compat.

Methods

impl TokenStream

fn empty() -> TokenStream

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_empty(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn concat<I: IntoIterator<Item=TokenStream>>(streams: I) -> TokenStream

🔬 This is a nightly-only experimental API. (rustc_private)

fn trees<'a>(&'a self) -> Cursor

🔬 This is a nightly-only experimental API. (rustc_private)

fn eq_unspanned(&self, other: &TokenStream) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Compares two TokenStreams, checking equality without regarding span information.