Module syntax::trees::ast::generic::def

Structs

Case

Cocase

Codata

Codef

Comatch

Ctor

Data

Decls

Def

Dtor

Impl

Match

Param

ParamInst

Instantiation of a previously declared parameter

Prg

Telescope

Wrapper type signifying the wrapped parameters have telescope semantics. I.e. each parameter binding in the parameter list is in scope for the following parameters.

TelescopeInst

Instantiation of a previously declared telescope

TypAbs

TypApp

Enums

Decl

Exp

Item

Type

Traits

HasPhase

Phase

Type Definitions

Args

Params