Trait syntax::visit::Visitor

pub trait Visitor<'ast>: Sized {
    fn visit_name(&mut self, _span: Span, _name: Name) { ... }

    fn visit_ident(&mut self, span: Span, ident: Ident) { ... }

    fn visit_mod(
        &mut self, 
        m: &'ast Mod, 
        _s: Span, 
        _attrs: &[Attribute], 
        _n: NodeId
    ) { ... }

    fn visit_foreign_item(&mut self, i: &'ast ForeignItem) { ... }

    fn visit_global_asm(&mut self, ga: &'ast GlobalAsm) { ... }

    fn visit_item(&mut self, i: &'ast Item) { ... }

    fn visit_local(&mut self, l: &'ast Local) { ... }

    fn visit_block(&mut self, b: &'ast Block) { ... }

    fn visit_stmt(&mut self, s: &'ast Stmt) { ... }

    fn visit_arm(&mut self, a: &'ast Arm) { ... }

    fn visit_pat(&mut self, p: &'ast Pat) { ... }

    fn visit_expr(&mut self, ex: &'ast Expr) { ... }

    fn visit_expr_post(&mut self, _ex: &'ast Expr) { ... }

    fn visit_ty(&mut self, t: &'ast Ty) { ... }

    fn visit_generics(&mut self, g: &'ast Generics) { ... }

    fn visit_where_predicate(&mut self, p: &'ast WherePredicate) { ... }

    fn visit_fn(
        &mut self, 
        fk: FnKind<'ast>, 
        fd: &'ast FnDecl, 
        s: Span, 
        _: NodeId
    ) { ... }

    fn visit_trait_item(&mut self, ti: &'ast TraitItem) { ... }

    fn visit_impl_item(&mut self, ii: &'ast ImplItem) { ... }

    fn visit_trait_ref(&mut self, t: &'ast TraitRef) { ... }

    fn visit_ty_param_bound(&mut self, bounds: &'ast TyParamBound) { ... }

    fn visit_poly_trait_ref(
        &mut self, 
        t: &'ast PolyTraitRef, 
        m: &'ast TraitBoundModifier
    ) { ... }

    fn visit_variant_data(
        &mut self, 
        s: &'ast VariantData, 
        _: Ident, 
        _: &'ast Generics, 
        _: NodeId, 
        _: Span
    ) { ... }

    fn visit_struct_field(&mut self, s: &'ast StructField) { ... }

    fn visit_enum_def(
        &mut self, 
        enum_definition: &'ast EnumDef, 
        generics: &'ast Generics, 
        item_id: NodeId, 
        _: Span
    ) { ... }

    fn visit_variant(
        &mut self, 
        v: &'ast Variant, 
        g: &'ast Generics, 
        item_id: NodeId
    ) { ... }

    fn visit_lifetime(&mut self, lifetime: &'ast Lifetime) { ... }

    fn visit_lifetime_def(&mut self, lifetime: &'ast LifetimeDef) { ... }

    fn visit_mac(&mut self, _mac: &'ast Mac) { ... }

    fn visit_mac_def(&mut self, _mac: &'ast MacroDef, _id: NodeId) { ... }

    fn visit_path(&mut self, path: &'ast Path, _id: NodeId) { ... }

    fn visit_path_list_item(
        &mut self, 
        prefix: &'ast Path, 
        item: &'ast PathListItem
    ) { ... }

    fn visit_path_segment(
        &mut self, 
        path_span: Span, 
        path_segment: &'ast PathSegment
    ) { ... }

    fn visit_path_parameters(
        &mut self, 
        path_span: Span, 
        path_parameters: &'ast PathParameters
    ) { ... }

    fn visit_assoc_type_binding(&mut self, type_binding: &'ast TypeBinding) { ... }

    fn visit_attribute(&mut self, attr: &'ast Attribute) { ... }

    fn visit_tt(&mut self, tt: TokenTree) { ... }

    fn visit_tts(&mut self, tts: TokenStream) { ... }

    fn visit_token(&mut self, _t: Token) { ... }

    fn visit_vis(&mut self, vis: &'ast Visibility) { ... }

    fn visit_fn_ret_ty(&mut self, ret_ty: &'ast FunctionRetTy) { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Each method of the Visitor trait is a hook to be potentially overridden. Each method's default implementation recursively visits the substructure of the input via the corresponding walk method; e.g. the visit_mod method by default calls visit::walk_mod.

If you want to ensure that your code handles every variant explicitly, you need to override each method. (And you also need to monitor future changes to Visitor in case a new method with a new default implementation gets introduced.)

Provided Methods

fn visit_name(&mut self, _span: Span, _name: Name)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_ident(&mut self, span: Span, ident: Ident)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_mod(
&mut self,
m: &'ast Mod,
_s: Span,
_attrs: &[Attribute],
_n: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_foreign_item(&mut self, i: &'ast ForeignItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_global_asm(&mut self, ga: &'ast GlobalAsm)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_item(&mut self, i: &'ast Item)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_local(&mut self, l: &'ast Local)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_block(&mut self, b: &'ast Block)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_stmt(&mut self, s: &'ast Stmt)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_arm(&mut self, a: &'ast Arm)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_pat(&mut self, p: &'ast Pat)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_expr(&mut self, ex: &'ast Expr)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_expr_post(&mut self, _ex: &'ast Expr)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_ty(&mut self, t: &'ast Ty)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_generics(&mut self, g: &'ast Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_where_predicate(&mut self, p: &'ast WherePredicate)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_fn(&mut self, fk: FnKind<'ast>, fd: &'ast FnDecl, s: Span, _: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_trait_item(&mut self, ti: &'ast TraitItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_impl_item(&mut self, ii: &'ast ImplItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_trait_ref(&mut self, t: &'ast TraitRef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_ty_param_bound(&mut self, bounds: &'ast TyParamBound)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_poly_trait_ref(
&mut self,
t: &'ast PolyTraitRef,
m: &'ast TraitBoundModifier
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_variant_data(
&mut self,
s: &'ast VariantData,
_: Ident,
_: &'ast Generics,
_: NodeId,
_: Span
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_struct_field(&mut self, s: &'ast StructField)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_enum_def(
&mut self,
enum_definition: &'ast EnumDef,
generics: &'ast Generics,
item_id: NodeId,
_: Span
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_variant(
&mut self,
v: &'ast Variant,
g: &'ast Generics,
item_id: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_lifetime(&mut self, lifetime: &'ast Lifetime)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_lifetime_def(&mut self, lifetime: &'ast LifetimeDef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_mac(&mut self, _mac: &'ast Mac)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_mac_def(&mut self, _mac: &'ast MacroDef, _id: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_path(&mut self, path: &'ast Path, _id: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_path_list_item(&mut self, prefix: &'ast Path, item: &'ast PathListItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_path_segment(
&mut self,
path_span: Span,
path_segment: &'ast PathSegment
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_path_parameters(
&mut self,
path_span: Span,
path_parameters: &'ast PathParameters
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_assoc_type_binding(&mut self, type_binding: &'ast TypeBinding)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_attribute(&mut self, attr: &'ast Attribute)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_tt(&mut self, tt: TokenTree)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_tts(&mut self, tts: TokenStream)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_token(&mut self, _t: Token)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_vis(&mut self, vis: &'ast Visibility)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_fn_ret_ty(&mut self, ret_ty: &'ast FunctionRetTy)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?